% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Components-class}
\alias{Components-class}
\title{An S4 class to specify the components of a Mix prior.}
\description{
An object of class \code{Components} is used to specify the
trend term in a \code{\link{Mix}} prior.
}
\section{Slots}{

\describe{
\item{\code{multVectorsMix}}{Multiplier applied to the default
value for \code{omegaVectorsMix}.}

\item{\code{nuVectorsMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaVectorsMix}.}

\item{\code{AVectorsMix}}{Scale for truncated half-t prior for
\code{omegaVectorsMix}.}

\item{\code{omegaVectorsMaxMix}}{Upper limit for truncated half-t prior
for \code{omegaVectorsMix}.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Components} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
An object of class \code{Components} is created
using function \code{\link{Components}}.
}
