% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Components}
\alias{Components}
\title{Specify priors for the components in a Mix prior.}
\usage{
Components(scale = HalfT())
}
\arguments{
\item{scale}{An object of class \code{\link{HalfT}}.}
}
\value{
An object of class \code{\linkS4class{Components}}.
}
\description{
In normal usage, it is better to accept the default prior
for the components than to specify it explicitly.
End users are therefore unlikely to need this function.
}
\details{
A \code{\link{Mix}} prior treats an interaction as a mixture of
normal distributions.  The means for these normal distributions
are formed by multiplying a component for each dimension
other than the 'along' dimension (typically the time dimension).
Each element of these components has a normal prior with mean
0 and standard deviation \code{scaleComponent}.  The prior
for \code{scaleComponent} can be set using the \code{scale}
argument.

Let \eqn{s} be the standard deviation of data \eqn{y},
or of \code{log(y)} in the case of a Poisson model without
exposure.  Let \eqn{m} be the \code{mult} argument. Then
the default value for scale parameter in the half-t distribution is

\tabular{ll}{
  \strong{Model} \tab \strong{Default} \cr
  Poisson with exposure  \tab \eqn{m 0.5} \cr
  Poisson without exposure \tab \eqn{ms 0.5} \cr
  binomial  \tab \eqn{m 0.5} \cr
  normal \tab \eqn{ms 0.5}
}
}
\examples{
Components()
## A tighter prior on 'scaleComponent'
Components(scale = HalfT(scale = 0.25))

}
\seealso{
Mixture priors are specified using \code{\link{Mix}}.
The weights in a mixture prior are specified using
\code{\link{Weights}}.
}
