% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Covariates-class}
\alias{Covariates-class}
\title{S4 class to specify the covariates term of a prior.}
\description{
An object of class \code{Covariates} holds the information required to
construct a covariates term in a prior for a main effect or interaction.
}
\section{Slots}{

\describe{
\item{\code{formula}}{An R \code{\link[stats]{formula}}.}

\item{\code{data}}{A data.frame holding the covariate data.}

\item{\code{contrastsArg}}{A named list of contrasts, as described in the
documentation for \code{\link[stats]{model.matrix}}.}

\item{\code{AEtaIntercept}}{Scale parameter for the normal
prior for the intercept.}

\item{\code{nuEtaCoef}}{Degrees of freedom for the t priors
for the coefficients.}

\item{\code{AEtaCoef}}{Scale parameter for the t priors
for the coefficients.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Covariates} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{Covariates} are created
by a call to function \code{\link{Covariates}}.
}
