% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{decomposition}
\alias{decomposition}
\alias{decomposition,Values-method}
\title{Decompose a Values array}
\usage{
decomposition(object)

\S4method{decomposition}{Values}(object)
}
\arguments{
\item{object}{An object of class \code{\link[dembase]{Values}}.}
}
\value{
A named list, the elements of which have class
\code{\link[dembase]{Values}}.
}
\description{
Decompose a \code{\link[dembase]{Values}} array into terms
made up of component dimensions.  Typically used to obtain
initial estimates of main effects and interactions,
as part of model building.
}
\details{
When building a Poisson model, the decomposition is usually
carried out on log-rates, and when building a binomial model,
it is usually carried out on logit-proportions.
}
\examples{
deaths <- Counts(demdata::VADeaths2)
popn <- Counts(demdata::VAPopn)
rates <- deaths/popn
log.rates <- log(rates)
ans <- decomposition(log.rates)
names(ans)
ans[1:3]
ans[["age:residence"]]
mean(log.rates)
round(sapply(ans, sum), 5)
all.equal(Reduce("+", ans), log.rates)
}
\references{
Chapter 12 of Bryant and Zhang,
\emph{Bayesian Demographic Estimation and Forecasting}.
}
