% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{Dispersion}
\alias{Dispersion}
\title{Specify the prior for the dispersion parameter in a CMP model.}
\usage{
Dispersion(mean = Norm(), scale = HalfT())
}
\arguments{
\item{mean}{An object of class \code{\linkS4class{Norm}}, specifying
the prior for the mean of the (logged) dispersion parameter.}

\item{scale}{An object of class \code{\linkS4class{HalfT}},
specifying the prior for the standard deviation of the (logged)
dispersion parameter.}
}
\value{
An object of class \code{\linkS4class{Dispersion}}.
}
\description{
Specify the prior for the dispersion parameter (often denoted
'nu') in a CMP or COM-Poisson model. The dispersion parameter
is assumed to have a distribution
\deqn{log \nu_i \sim N(\mu_i, \sigma^2)}
\deqn{\mu_i \sim N(M, S^2)}
\deqn{\sigma \sim trunc-half-t(df, A, max)}
}
\details{
The \code{mean} argument is used to specify the second equation,
and the \code{scale} argument is used to specify the third.
\code{M} in the second equation defaults to 0, and \code{S}
defaults to 1. \code{df} in the third equation defaults to 7,
and \code{A} defaults to 1.
}
\examples{
Dispersion()
Dispersion(mean = Norm(mean = -1, sd = 0.1))
Dispersion(mean = Norm(mean = -1, sd = 0.1),
           scale = HalfT(scale = 0.1))
}
\seealso{
CMP models are specified with function \code{\link{CMP}}.
}
