% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{DLM}
\alias{DLM}
\title{Specify a Dynamic Linear Model (DLM) prior.}
\usage{
DLM(along = NULL, level = Level(), trend = Trend(), damp = Damp(),
  season = NULL, covariates = NULL, error = Error())
}
\arguments{
\item{along}{Name of a dimension.}

\item{level}{An object of class \code{\linkS4class{Level}}
specifying the prior for \code{errorLevel}.}

\item{trend}{An object of class \code{\linkS4class{Trend}}
(the default) or \code{NULL}.  If \code{trend} is \code{NULL},
then \code{DLM} produces local level model; otherwise
\code{DLM} produces a linear trend model.}

\item{damp}{An object of class \code{\linkS4class{Damp}}
(the default) or \code{NULL}.  If \code{damp} is \code{NULL}
the level or trend terms are not damped.}

\item{season}{An object of class \code{\linkS4class{Season}}
or \code{NULL} (the default).  If \code{season} is
\code{NULL}, the prior does not include a seasonal effect.}

\item{covariates}{An object of class
\code{\linkS4class{Covariates}} or \code{NULL}
(the default).  If \code{covariates} is
\code{NULL}, the prior does not include covariates.}

\item{error}{An object of class \code{\link{Error}}.}
}
\value{
An object of class \code{\linkS4class{SpecDLM}}.
}
\description{
Specify a prior for main effects or interactions in which units that are
neighbours are more strongly correlated than units that are distant
from each other. Dynamic linear models were developed for time
series, but are often appropriate for non-time dimensions, such as age.
}
\details{
When applied to a main effect, the prior specified by \code{DLM} has the
form

\code{parameter[j] = level[j] + season[j] + covariates[j] + error[j].}

When applied to an interaction, a DLM prior contains a separate series
along one dimension (typically a time dimension) for each combination
of values for the remaining series.  Using \code{k} to index location
on the "along" dimension, and \code{l} to index location on the remaining
dimensions, the prior has the form

\code{
parameter[k,l] = level[k,l] + season[k,l] + covariates[k,l] + error[k,l].
}

For instance, if \code{parameter} is a region:time interaction involving 10
regions and 30 periods, and the "along" dimension is time, then
\code{k=1,...,30} and \code{l=1,...,10}.  If \code{parameters} is an
age:sex:region interaction involving 101 age groups, 2 sexes, and 10
regions, and the "along" dimension is age, then \code{k=1,...,101} and
\code{l = 1,...,20}.

Season and covariate terms are specified using functions \code{\link{Season}}
and \code{\link{Covariates}}.  Both are optional.  Error terms are 
specified using  function \code{\link{Error}}.
}
\section{The "along" dimension}{


If the "along" dimension is not specified, then \code{DLM} looks
for a dimension with \code{\link[dembase]{dimtype}}
\code{"time"}.  If it does not find one, it looks for a dimension with
dimtype \code{"age"}.

A dimension with dimtype \code{"state"} can be used as an \code{"along"}
dimension.  Doing so probably make senses if the dimension measures
a qualitative variable such as income or years of schooling.  It probably
does not makes sense if the dimension measures a qualitative variable
such as region or occupation.
}

\section{The local level model}{


In a local level model, the level term follows a random walk.  With
a main effect,

\code{level[j] = damp * level[j-1] + errorLevel[j],}

and with an interaction,

\code{level[k,l] = damp * level[k-1,l] + errorLevel[k,l],}

where \code{0 < damp <= 1}.  The documentaion for function
\code{\link{Damp}} describes the role played by the \code{damp} term.
}

\section{The linear trend model}{


In a linear level model, the level term follows a random walk but with
an underlying trend.  The trend term itself follows a random walk, implying
that it can change magnitude or direction.  With a main effect,

\code{level[j] = level[j-1] + trend[j-1] + errorLevel[j]}
\code{trend[j] = trend[j-1] + errorTrend[j],}

and with an interaction,

\code{level[k,l] = level[k-1,l] + trend[k-1,l] + errorLevel[k,l]}
\code{trend[k,l] = damp * trend[k-1,l] + errorTrend[k,l].}

where \code{0 < damp <= 1}.  The documentaion for function
\code{\link{Damp}} describe the role played by the \code{damp} term.
}

\section{Error terms}{


The error term for the level, \code{errorLevel}, has the form

\code{errorLevel[j] ~ N(0, scaleLevel^2)}

or

\code{errorLevel[k,l] ~ N(0, scaleLevel^2)}.

Similarly, the error term for the trend, \code{errorTrend}, has the form

\code{errorLevel[j] ~ N(0, scaleLevel^2)}

or

\code{errorLevel[k,l] ~ N(0, scaleLevel^2)}.

\code{scaleLevel} and \code{scaleTrend} both have half-t priors.
Non-default versions of these can be specified using function 
\code{\link{HalfT}}.
}

\examples{
## damped linear trend model with no seasonal effect
## and no covariates
DLM()

## damped local level model with no seasonal effect
## and no covariates
DLM(trend = NULL)

## local level model with no damping, no seasonal effect,
## and no covariates
DLM(trend = NULL, damp = NULL)

## damped linear trend model with seasonal effect
## and no covariates
DLM(season = Season(n = 4))

## linear trend model with covariates, but no damping
## and no seasonal effects
gdp.data <- data.frame(year = 2000:2016,
                       gdp = rnorm(17))
gdp.covariates <- Covariates(mean ~ gdp, data = gdp.data)
DLM(damp = NULL, covariates = gdp.covariates)

## robustified damped linear trend model
DLM(error = Error(robust = TRUE))
}
\references{
Petris, G., Petrone, S., & Campagnoli, P. (2009). \emph{Dynamic Linear
Models with R}. Springer.

Prado, R., & West, M. (2010). \emph{Time Series: Modeling, Computation,
and Inference}. CRC Press.
}
\seealso{
\code{\link{ExchFixed}} and \code{\link{Exch}} specify exchangeable
in which the order of the units conveys no information about the
correlations between them.  \code{DLM}, \code{\link{ExchFixed}},
and \code{\link{Exch}} are usually called as part of a call
to function \code{\link{Model}}.
}
