% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{DLMS}
\alias{DLMS}
\title{Short version of 'DLM' function for creating Dynamic Linear Model priors}
\usage{
DLMS(l = TRUE, t = TRUE, d = TRUE, scale)
}
\arguments{
\item{l}{Logical. If \code{TRUE} (the default), the model includes
a level term.}

\item{t}{Logical. If \code{TRUE} (the default), the model includes
a trend term.}

\item{d}{Logical. If \code{TRUE} (the default), the model is
damped, using the defaults for \code{\link{Damp}}.}

\item{scale}{A positive number. The value of the \code{scale}
parameter to be used in the priors for standard
deviations in level, trend, and error terms.}
}
\value{
An object of class \code{\linkS4class{SpecDLM}}.
}
\description{
WARNING - THIS FUNCTION IS EXPERIMENTAL.
THE INTERFACE IS NOT YET FINALISED.
}
\examples{
DLMS(scale = 0.1)
## identical to
DLM(level = Level(scale = HalfT(scale = 0.1)),
    trend = Trend(scale = HalfT(scale = 0.1)),
    error = Error(scale = HalfT(scale = 0.1)))

DLMS(t = FALSE, d = FALSE, scale = 0.1)
## identical to
DLM(level = Level(scale = HalfT(scale = 0.1)),
    trend = NULL,
    damp = NULL,
    error = Error(scale = HalfT(scale = 0.1)))
}
