% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Error-class}
\alias{Error-class}
\alias{ErrorNorm-class}
\alias{ErrorRobust-class}
\title{S4 classes to specify error terms for priors.}
\description{
An object of class \code{Error} holds the information required to construct
an error term in a prior for a main effect or interaction.  Errors can have
a normal or \emph{t} distributions.  Error terms with normal distributions are 
described by objects of class \code{ErrorNorm}, and error terms with \emph{t}
distributions are described by objects of class \code{ErrorRobust}.
}
\details{
\code{ErrorNorm} and \code{ErrorRobust} are both subclasses of virtual
class \code{Error}.

The \code{scale} parameters for the normal and t-distribution priors
have truncated half-t priors.
}
\section{Slots}{

\describe{
\item{\code{nuBeta}}{Degrees of freedom for the \emph{t} distribution. (\code{ErrorRobust} only).}

\item{\code{nuTau}}{Degrees of freedom for the prior for the \code{scale} parameter.}

\item{\code{ATau}}{Scale for the prior for the \code{scale} parameter.}

\item{\code{tauMax}}{Maximum value for the \code{scale} parameter.}
}}
\seealso{
Objects of class \code{Error} are created by a call to function
\code{\link{Error}}.
}

