% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Error}
\alias{Error}
\title{Specify the error term in a prior for a main effect or interaction.}
\usage{
Error(robust = FALSE, df = 4, scale = HalfT())
}
\arguments{
\item{robust}{Whether to use a \emph{t} distribution.  Defaults to
\code{FALSE}.}

\item{df}{The degrees of freedom for the \emph{t} distribution.
Defaults to 4.}

\item{scale}{Prior for the \code{scale} parameter.  An object of class
\code{\linkS4class{HalfT}}.}
}
\value{
An object of class \code{\linkS4class{Error}}
}
\description{
Priors for main effects or interactions, such as age effects or
age-sex interactions, typically include an error term.  The error
terms can be specified using  function \code{Error}.
}
\details{
If \code{robust = FALSE} (the default), then the error term has a normal
distribution,

\code{error[j] ~ N(0, scale^2)}.

The \code{scale} parameter has a half-\emph{t} prior, which is described in the
documentation for function \code{\link{HalfT}}.

If \code{robust = TRUE}, then, by default, the error term has a \emph{t}
distribution with 4 degrees of freedom,

\code{error[j] ~ t(4, 0, scale^2)}.

Users can supply alternative values for the degrees of freedom.
The \code{scale} parameter has a half-\emph{t} prior, described
in the documentation for \code{\link{HalfT}}.
}
\examples{
Error()
Error(robust = TRUE)
Error(robust = TRUE, df = 10)
Error(scale = HalfT(scale = 0.5, max = 1))
Error(scale = HalfT(mult = 0.5))
}

