% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-functions.R
\name{estimateAccount}
\alias{estimateAccount}
\title{Estimate demographic account and models from multiple noisy datasets.}
\usage{
estimateAccount(
  account,
  systemModels,
  datasets,
  dataModels,
  concordances = list(),
  weights = list(),
  dominant = c("Female", "Male"),
  updateInitialPopn = TRUE,
  usePriorPopn = TRUE,
  scaleNoise = 0,
  filename = NULL,
  nBurnin = 1000,
  nSim = 1000,
  nChain = 4,
  nThin = 1,
  parallel = TRUE,
  nCore = NULL,
  outfile = NULL,
  nUpdateMax = 50,
  verbose = FALSE,
  useC = TRUE
)
}
\arguments{
\item{account}{An object of class
\code{\link[dembase:DemographicAccount-class]{DemographicAccount}},
giving the initial values for the estimation of the account.}

\item{systemModels}{A list of objects of class
\code{\linkS4class{SpecModel}}
specifying models for the demographic series.}

\item{datasets}{A named list of objects of class
\code{\link[dembase:DemographicArray-class]{Counts}}.}

\item{dataModels}{A list of objects of class
\code{\linkS4class{SpecModel}} specifying models for the datasets.}

\item{concordances}{A named list of
\code{\link[dembase:Concordance-class]{concordances}},
which are applied to the series in the account before they are
supplied to the corresponding data model.}

\item{weights}{A named list of
\code{\link[dembase:Counts-class]{Counts}} objects,
providing weights for any Normal models among the system models.}

\item{dominant}{Either \code{"Female"} (the default) or \code{"Male"}.
Determines which sex is used to generate exposures in the system
model for births.}

\item{updateInitialPopn}{If \code{TRUE} (the default) population counts
in the first year of the account are inferred as part of
the overall estimation process. If \code{FALSE}, the values supplied
for the first year are treated as error-free and never updated.}

\item{usePriorPopn}{Whether to take account of the prior model
for population when inferring values for the account.
Defaults to \code{TRUE}.}

\item{scaleNoise}{Governs noise added to Metropolis-Hastings
ratio. Should be non-zero only when trying to generate
initial values. Currently experimental, and may change.}

\item{filename}{The name of a file where output is collected.}

\item{nBurnin}{Number of iteration discarded before recording begins.}

\item{nSim}{Number of iterations carried out during recording.}

\item{nChain}{Number of independent chains to use.}

\item{nThin}{Thinning interval.}

\item{parallel}{Logical.  If \code{TRUE} (the default), parallel processing
is used.}

\item{nCore}{The number of cores to use, when \code{parallel}
is \code{TRUE}.  If no value supplied, defaults to \code{nChain}.}

\item{outfile}{Where to direct the ‘stdout’ and ‘stderr’ connection
output from the workers when parallel processing.  Passed to function
\code{[parallel]{makeCluster}}.}

\item{nUpdateMax}{Maximum number of iterations completed before releasing
memory.  If running out of memory, setting a lower value than the default
may help.}

\item{verbose}{Logical.  If \code{TRUE} (the default) a message is
printed at the end of the calculations.}

\item{useC}{Logical.  If \code{TRUE} (the default), the calculations
are done in C.  Setting \code{useC} to \code{FALSE} may be useful
for debugging.}
}
\description{
Infer the contents of a demographic account, and fit models
describing series within the account, using multiple noisy datasets.
\emph{The function is still under construction.}
}
\references{
Bryant, J., Graham, P. Bayesian demographic accounts:
Subnational population estimation using multiple datasources. 2013.
\emph{Bayesian Analysis}
}
\seealso{
\code{\link{estimateModel}}, \code{\link{estimateCounts}}
}
