% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{Exact}
\alias{Exact}
\title{Specify that a dataset is to be treated as error-free.}
\usage{
Exact()
}
\value{
An object of class \code{\linkS4class{SpecLikelihood}}.
}
\description{
Specify that a dataset gives complete, and error-free,
measurements of a series in a demographic account.
}
\details{
The dataset is used to populate the corresponding
series in the demographic account, and after that
the series is never updated.
}
\examples{
Exact()
}
\seealso{
\code{Exact} is typically used as
part of a call to function \code{\link{Model}}.
}
