% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Exch}
\alias{Exch}
\title{Specify an exchangeable prior.}
\usage{
Exch(covariates = NULL, error = Error())
}
\arguments{
\item{covariates}{An object of class
\code{\linkS4class{Covariates}} or \code{NULL}
(the default).  If \code{covariates} is
\code{NULL}, the prior does not include covariates.}

\item{error}{An object of class \code{\link{Error}}.}
}
\value{
An object of class \code{\linkS4class{SpecExch}}.
}
\description{
Specify a prior for main effects or interactions in which the units are
exchangeable.  Units are exchangeable if the labels or ordering of the
units does not provide any information about their distribution.
}
\details{
The prior specified by \code{Exch} has the form

\code{parameter[j] = covariates[j] + error[j].}

Including a covariates term can make the assumption of exchangeability
more plausible.  For instance, treating regional effects as exchangeable
is more plausible after the region effects have been adjusted for
differences in regions' incomes and population sizes.  A covariates term
can be specified using function \code{\link{Covariates}}.

Errors can have a normal or \emph{t} distribution.  The error term is
specified using function \code{\link{Error}}.
}
\examples{
Exch()
Exch(error = Error(robust = TRUE))
Exch(error = Error(scale = HalfT(scale = 0.5)))

## with covariates
data <- data.frame(region = letters[1:10],
                   income = rnorm(10))
Exch(covariates = Covariates(mean ~ income, data = data))
}
\seealso{
\code{\link{ExchFixed}} specifies a simpler
exchangeable prior.  \code{\link{DLM}} specifies a
non-exchangeable prior.
}

