% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{ExchFixed}
\alias{ExchFixed}
\title{Specify a simple exchangeable prior.}
\usage{
ExchFixed(sd = NULL, mult = 1)
}
\arguments{
\item{sd}{Standard deviation.  Optional.}

\item{mult}{Multiplier applied to \code{sd}, if \code{sd}
is generated automatically.  Defaults to 1.}
}
\value{
An object of class \code{\linkS4class{SpecExchFixed}}.
}
\description{
A simpler exchangeable prior than \code{\link{Exch}}.  The units are
assumed to be drawn from a normal distribution with no covariates and
a known standard deviation.
}
\details{
The prior specified by \code{ExchFixed} has the form

\code{parameter[j] ~ N(0, sd^2).}

If a value for \code{sd} is not supplied by the user, then a value is
generated when function \code{\link{estimateModel}},
\code{\link{estimateCounts}}, or \code{\link{estimateAccount}} is called.
Let \eqn{s} be the standard deviation of data \code{y}, or of \code{log(y)}
in the case of a Poisson model without exposure, and let \eqn{m} be
the \code{mult} argument.  The default value for \code{sd} is then

\tabular{ll}{
  \strong{Model} \tab \strong{Default} \cr
  Poisson with exposure \tab \eqn{m} \cr
  Poisson without exposure \tab  \eqn{ms} \cr
  binomial \tab \eqn{m} \cr
  normal \tab  \eqn{ms}.
}

These value are designed to be weakly informative, in that they favour
values for \code{sd} that are demographically plausible.

\code{ExchFixed} can be useful a dimension with few elements, such as sex
dimension, where there is insufficient information to justify more
complicated priors.
}
\examples{
## default standard deviation
ExchFixed()
## user-specified standard deviation
ExchFixed(sd = 0.5)
## reduce the size of the automatically-generated 'sd'
ExchFixed(mult = 0.5)
## increase the size of the automatically-generated 'sd'
ExchFixed(mult = 2)
}
\seealso{
\code{\link{Exch}} specifies a more complicated
exchangeable prior.
}

