% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-functions.R
\name{fetch}
\alias{fetch}
\title{Extract estimates from model output.}
\usage{
fetch(filename, where = character(), iterations = NULL, impute = FALSE)
}
\arguments{
\item{filename}{The filename used by the estimate function.}

\item{where}{A character vector describing the path to the item to be
extracted.}

\item{iterations}{A vector of positive integers giving the iterations to be
extracted if an item has multiple iterations.}

\item{impute}{Logical. Whether to impute missing values.}
}
\value{
Parameters that were estimated from the data typically have class
\code{DemographicArray} and have a dimension with
\code{\link[dembase]{dimtype}} \code{"iteration"}.  Other elements
stored in \code{object} have a variety of classes.
}
\description{
Functions \code{\link{estimateModel}}, \code{\link{estimateCounts}},
and \code{\link{estimateAccount}}, and their \code{predict} counterparts,
send their output to a simple database specified by \code{filename}.
Function \code{fetch} extracts estimates from this database.
}
\details{
Estimates are stored in a hierarchical structure.  The structure be viewed
with function \code{\link{listContents}}.  Function \code{fetch} uses
a description of the path through the hierarchical structure to locate
estimates.  The \code{where} argument to \code{fetch} is a character vector,
giving names of nodes in the hierarchy.  Partial matching is used with
the names of the nodes, so names can be shortened, provided they are still
long enough to uniquely identify a node.  (To make code self-documenting,
it is still best to use the full names in production code.)

Using the \code{iterations} to select a subset of iterations can be useful
if the object being extracted is large relative to memory or if calculations
are running slowly.

Where possible, \code{\link{estimateModel}}, \code{\link{estimateCounts}},
and \code{\link{estimateAccount}} avoid imputing missing values during
model fitting, since the imputed values are typically highly correlated
with other unknown quantities, which can slow convergence.  If a batch
of estimates has missing values and if \code{impute} is \code{TRUE} the
missing values will be imputed at the time they are fetched.
}
\examples{
deaths <- demdata::VADeaths2
popn <- demdata::VAPopn
deaths <- round(deaths)
deaths <- Counts(deaths)
popn <- Counts(popn)
filename <- tempfile()
estimateModel(Model(y ~ Poisson(mean ~ age * sex)),
              y = deaths,
              exposure = popn,
              filename = filename,
              nBurnin = 20,
              nSim = 20,
              nChain = 2,
              parallel = FALSE)
## the model hasn't been run for long enough,
## but we'll extract some results anyway

## examine structure
listContents(filename)

## extract means from likelihood model
rate <- fetch(filename,
              where = c("model", "likelihood", "rate"))
plot(rate)

## only supply enough of component of 'where' to identify
rate <- fetch(filename,
              where = c("mo", "l", "r"))

## extract all iterations of intercept
fetch(filename,
      where = c("model", "prior", "(Intercept)"))

## extract every fifth iteration
fetch(filename,
      where = c("model", "prior", "(Intercept)"),
      iterations = seq(from = 5, by = 5, to = 40))
}
\references{
The weak identification of main effects and interactions is discussed in

Gelman, A. (2005). Analysis of variance: Why it is more important than ever.
\emph{Annals of Statistics}, 33(1):1-53.

and

Nelder, J. A. (1994). The statistics of linear models: Back to basics.
\emph{Statistics and Computing}. 4: 221-234.
}
\seealso{
\code{fetch} is used to extract output from functions 
\code{\link{estimateModel}}, \code{\link{estimateCounts}},
and \code{\link{estimateAccount}}.  Function \code{\link{listContents}}
shows the internal structure of the output, which is useful for
constructing the \code{where} argument to \code{fetch}.
}
