% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-functions.R
\name{fetchBoth}
\alias{fetchBoth}
\title{Extract combined results from estimation and prediction.}
\usage{
fetchBoth(filenameEst, filenamePred, where, iterations = NULL,
  normalize = TRUE, impute = FALSE)
}
\arguments{
\item{filenameEst}{The filename used for estimation.}

\item{filenamePred}{The filename used for prediction.}

\item{where}{A character vector describing the path to the item to be
extracted.}

\item{iterations}{A vector of positive integers giving the iterations to be
extracted if an item has multiple iterations.}

\item{normalize}{Logical. Whether to centre estimates of main effects and
interactions from hierarchical models.  Defaults to \code{TRUE}.}

\item{impute}{Logical. Whether to impute missing values.}
}
\description{
Combined results from calling \code{\link{estimateModel}} followed
by \code{\link{predictModel}}, \code{\link{estimateCounts}} followed
by \code{\link{predictCounts}}, or \code{\link{estimateAccount}}
followed by \code{\link{predictAccount}}.  (Note - functions
predictCounts, estimateAccount, and predictAccount have not been
written yet.)
}
\details{
Identical results can be obtained by calling \code{\link{fetch}}
separately on the files created by the estimation and prediction functions,
and then combining.  However, \code{fetchBoth} is more convenient.

\code{fetchBoth} is particularly convenient in the case of main effects,
since it normalises the combines the estimates and predictions before
normalising, rather than normalising each separately.  (See the
documentation for \code{\link{fetch}} for a discussion of normalisation.)
}
\examples{
\dontrun{
deaths <- demdata::us.deaths
exposure <- demdata::us.exposure
deaths <- Counts(deaths,
                 dimscales = c(year = "Intervals"))
exposure <- Counts(exposure,
                   dimscales = c(year = "Intervals"))
filename.est <- tempfile()
filename.pred <- tempfile()
model <- Model(y ~ Poisson(mean ~ age + sex + year))
estimateModel(model,
              filename = filename.est,
              y = deaths,
              exposure = exposure,
              nBurnin = 5,
              nSim = 5,
              nChain = 2,
              parallel = FALSE)
predictModel(filenameEst = filename.est,
             filenamePred = filename.pred,
             n = 5)
## the model is too simple, and hasn't been run for long
## enough, but we'll extract some results anyway...
rate <- fetchBoth(filenameEst = filename.est,
                  filenamePred = filename.pred,
                  where = c("model", "likelihood", "rate"))
year <- fetchBoth(filenameEst = filename.est,
                  filenamePred = filename.pred,
                  where = c("model", "prior", "year"))
}
}
\seealso{
Function \code{\link{fetch}} extracts results from a single file.
}

