% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-functions.R
\name{fetchFiniteSD}
\alias{fetchFiniteSD}
\title{Finite-population standard deviations.}
\usage{
fetchFiniteSD(filename, probs = c(0.025, 0.5, 0.975), iterations = NULL)
}
\arguments{
\item{filename}{The name of the file where the output from the
\code{estimate} function is kept.}

\item{probs}{A vectors of quantiles, for summarizing the standard deviations.}

\item{iterations}{A vector of positive integers giving the iterations to be
extracted.}
}
\value{
A list of objects of class \code{"FiniteSD"} if
\code{object} contains several hierarchical models; a single
\code{"FiniteSD"} object if' \code{object} contains
a single hierchical model; and \code{NULL} if \code{object}
contains no hierarchical models.
}
\description{
Extract the finite-population standard deviation of main effects
and interactions in hierarchical models.
}
\details{
In the hierarchical models discussed in \code{\link{Model}}, cell means
are moderned using terms formed from cross-classifying dimensions such as
age or sex, or from interactions between these dimensions.  Adapting ideas
from ANOVA, Gelman (2004: 408-409) propose that the finite-population
standard deviations of these terms be used as a measure of the terms'
relative importance.  For instance, is an 'age' term has a greater
finite-population standard deviation than a 'region' term in a model, this
suggests that age is a more important predictor of outcomes than region.
}
\references{
Gelman, A. (2004) Analysis of variance:
why it is more important than ever
(with discussion). \emph{Annals of Statistics}. 33(1): 1-53.
}
\seealso{
\code{\link{estimateModel}}, \code{\link{estimateCounts}},
\code{\link{estimateAccount}}
}
