% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-21-Summary.R,
%   R/DemographicArray-methods.R
\docType{class}
\name{FiniteSD-class}
\alias{FiniteSD-class}
\alias{show,FiniteSD-method}
\title{S4 class to hold finite-population standard deviations.}
\usage{
\S4method{show}{FiniteSD}(object)
}
\arguments{
\item{object}{Object of class \code{FiniteSD}.}
}
\description{
Object of class \code{FiniteSD} hold the finite-population standard
deviations of main effects and interactions from hierarchical models.
See the documentation for \code{\link{fetchFiniteSD}} for details.
}
\section{Slots}{

\describe{
\item{\code{df}}{Integer vector holding the degrees of freedom.}
}}

