% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{gelmanDiag}
\alias{gelmanDiag}
\title{Obtain potential scale reduction factors (Rhats).}
\usage{
gelmanDiag(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{SummaryResults}}.}
}
\value{
A named numeric vector.
}
\description{
Extract potential scale reduction factors (Rhats) from an object of
class \code{\linkS4class{SummaryResults}}. See the documentation for
\code{\link{fetchSummary}} for details.
}
\examples{
library(demdata)
deaths <- Counts(round(VADeaths2))
popn <- Counts(VAPopn)
filename <- tempfile()
estimateModel(Model(y ~ Poisson(mean ~ sex)),
              y = deaths,
              exposure = popn,
              filename = filename,
              nBurnin = 2,
              nSim = 20,
              nChain = 2,
              parallel = FALSE)
summary.deaths <- fetchSummary(filename)
gelmanDiag(summary.deaths)
}
\seealso{
\code{\link{fetchSummary}}
}
