% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{halft-distn}
\alias{halft-distn}
\alias{rhalft,}
\alias{qhalft,}
\alias{phalft,}
\alias{dhalft}
\alias{rhalft}
\alias{qhalft}
\alias{phalft}
\title{The half-t distribution.}
\usage{
rhalft(n, df, scale = 1)

qhalft(p, df, scale = 1)

phalft(q, df, scale = 1)

dhalft(x, df, scale = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{df}{Degrees of freedom. Positive. Can be non-integer,
and can be \code{Inf}.}

\item{scale}{Dispersion parameter.}

\item{p}{Vector of quantiles.}

\item{q}{Vector of probabilities.}

\item{x}{Vector of quantiles.}
}
\value{
\code{dhalft} gives the density, \code{phalft} gives
    the distribution function, \code{qhalft} gives the
    quantile function, and \code{rhalft} generates random
    deviates.
}
\description{
Density, distribution function, quantile function and random
generation for the halt-t distribution.
}
\details{
The half-t distribution is also known as the folded-t distribution.

If \eqn{X} has a \emph{t} distribution with degrees of freedom
\eqn{v}, location 0, and scale \code{s}, then \eqn{|X|}
has a half-\emph{t} distribution with degrees of freedom \eqn{v}
and scale \code{s}.

Internally, the functions all call the corresponding functions
for the \code{\link[=TDist]{t distribution}}.
}
\examples{
dhalft(x = 0.5, df = 7, scale = 0.5)
qhalft(p = 0.9, df = 4)
phalft(q = 0.5, df = 7, scale = 2)
rhalft(n = 5, df = 30)
}
\references{
Based on Brazauskas, V., and Kleefeld, A. (2011) Folded and log-folded-t
distributions as models for insurance loss data.
\emph{Scandinavian Actuarial Journal} 59-74.
}
\seealso{
Function \code{\link{plotHalfT}} plots density and distribution
functions for half-\emph{t} distributions.
}
