% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{HalfT}
\alias{HalfT}
\title{Specify a half-t distribution.}
\usage{
HalfT(df = 7, scale = NULL, max = NULL, mult = 1)
}
\arguments{
\item{df}{Degrees of freedom of the half-\emph{t} distribution.
A positive number, defaulting to 7}

\item{scale}{Scale parameter for the  half-\emph{t} distribution.
A positive number.}

\item{max}{A positive number.  If finite, the half-\emph{t}
distribution is truncated at this point.}

\item{mult}{Multiplier applied to \code{scale} if \code{scale}
is generated automatically.  Defaults to 1.}
}
\value{
Object of class \code{\linkS4class{HalfT}}.
}
\description{
If \code{x} has a \emph{t} distribution, then \code{abs(x)} has a
half-\emph{t} distribution, also known as a folder \emph{t}
distribution.  In package \pkg{demest}, most standard deviation or
scale parameters have half-\emph{t} priors.
}
\details{
A half-\emph{t} distribution with degrees of freedom \eqn{n} and
scale \eqn{A} has density

\deqn{p(y) \propto ((y^2)/n + A^2)^(-(n+1)/2).}

Setting a maximum value that the standard deviation or scale
parameter can take (and so specifying a \emph{truncated} half-\emph{t}
distribution) can reduce numerical problems, and dramatically
improve convergence.  By default, the maximum value is set to the 0.999
quantile.Users can specify alternative values, including infinity.

\code{HalfT} is typically used to specify the prior for
a main effect or interaction.  In this case, if a value for \code{scale}
is not specified, a default value is determined when function
\code{\link{estimateModel}}, \code{\link{estimateCounts}}, or
\code{\link{estimateAccount}} is called. Let \eqn{s} be the standard
deviation of data \eqn{y}, or of \code{log(y)} in the case of a
Poisson model without exposure. Let \eqn{d} be the degree of an
interaction: for instance, an interaction between age and sex has
degree 2, and an interaction between age, sex, and region has degree 3.
Let \eqn{m} be the \code{mult} argument. The default value for \code{scale}
is then

\tabular{lll}{
  \strong{Model} \tab \strong{Term} \tab \strong{Default} \cr
  Poisson with exposure \tab Main effect \tab \eqn{m} \cr 
  Poisson with exposure \tab Interaction of degree \eqn{d} \tab \eqn{m 0.5^{d-1}} \cr
  Poisson without exposure \tab Main effect \ \eqn{ms} \cr
  Poisson without exposure \tab Interaction of degree \eqn{d} \tab \eqn{ms 0.5^{d-1}} \cr
  binomial \tab Main effect \tab \eqn{m} \cr
  binomial \tab Interaction of degree \eqn{d} \tab \eqn{m 0.5^{d-1}} \cr
  normal \tab Main effect \tab \eqn{ms} \cr
  normal \tab Interaction of degree \eqn{d} \tab \eqn{ms 0.5^{d-1}}.
}
}
\examples{
HalfT()
HalfT(scale = 0.5)
HalfT(df = 1, scale = 10)
HalfT(scale = 0.5, max = 4)
HalfT(mult = 0.5)
HalfT(mult = 2)
}
\references{
Brazauskas, V., and Kleefeld, A. (2011) Folded and log-folded-t
distributions as models for insurance loss data.
\emph{Scandinavian Actuarial Journal} 59-74.

Gelman, A. (2006). Prior distributions for variance parameters in
hierarchical models (comment on article by Browne and Draper).
\emph{Bayesian Analysis} 1, 515-534.

Gelman, A., Carlin, J.B., Stern, H.S. and Rubin, D.B., 2014.
\emph{Bayesian Ddata Analysis. Third Edition.} Boca Raton, FL, USA:
Chapman & Hall/CRC. Section 16.3.
}
\seealso{
\code{\link{Error}}, \code{\link{Exch}}, \code{\link{Exch}}
}

