% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Initial-class}
\alias{Initial-class}
\title{An S4 class to specify a normal prior for a scalar parameter.}
\description{
An object of class \code{Initial} is used to specify the
prior for the initial trend term, typically as part of a
call to function \code{\link{DLM}}.
}
\section{Slots}{

\describe{
\item{\code{A}}{Scale.}

\item{\code{mean}}{The mean of the normal distribution.  Defaults to 0.}

\item{\code{mult}}{Multiplier applied to default value for \code{A}.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of an
\code{Initial} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Objects of class \code{Initial} are created
by calls to function \code{\link{Initial}}.
}

