% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Initial}
\alias{Initial}
\title{Specify the prior for the initial value of the trend term in a DLM prior.}
\usage{
Initial(mean = 0, sd = NULL, mult = 1)
}
\arguments{
\item{mean}{The mean of the prior distribution.  Defaults to 0.}

\item{sd}{The standard deviation of the prior distribution.}

\item{mult}{Multiplier applied to \code{scale} if \code{scale}
is generated automatically.  Defaults to 1.}
}
\value{
An object of class \code{\linkS4class{Initial}}.
}
\description{
Specify the prior for the initial value of the trend term, typically as
part of a call to function \code{\link{Trend}}.
}
\details{
In the linear trend version of a DLM prior, the trend term has the form

\code{trend[0] ~ N(mean, sd^2)}
\code{trend[j] ~ damp * trend[j-1] + errorLevel[j], j = 1,...,J}.

By default, the mean in the prior for \code{trend[0]} defaults to 0.
The default for the \code{sd} is set using the same rules as the
default for the scale term in a \code{\link{HalfT}} prior.
}
\examples{
Initial()
Initial(mean = 0.1, sd = 0.3)
Initial(mean = 0.02, mult = 0.1)
}
\seealso{
\code{Initial} is usually called as part of a call to function
\code{\link{Trend}}.
}

