% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{InvChiSq-class}
\alias{InvChiSq-class}
\title{An S4 class to specify a truncated scaled
inverse-chi-squared distribution.}
\description{
Truncated scaled inverse-chi-squared distributions
}
\section{Slots}{

\describe{
\item{\code{nu}}{Degrees of freedom.}

\item{\code{A}}{Scale.}

\item{\code{scaleMax}}{Maximum value.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{InvChiSq} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{InvChiSq} are created
by calls to function \code{\link{InvChiSq}}.
}
