% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{InvChiSq}
\alias{InvChiSq}
\title{Specify an inverse chi-squared distribution.}
\usage{
InvChiSq(df, scaleSq, max = NULL)
}
\arguments{
\item{df}{Degrees of freedom. A positive number.}

\item{scaleSq}{Scale parameter. A positive number.}

\item{max}{A positive number.  If finite, the
distribution is truncated at this point.}
}
\value{
Object of class \code{\linkS4class{InvChiSq}}.
}
\description{
Specify an inverse chi-squared distribution,
for use as a prior..
}
\examples{
InvChiSq(df = 10, scale = 10, max)
}
