% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Known}
\alias{Known}
\title{Specify a prior where the mean varies but is treated as known.}
\usage{
Known(mean, sd = 0)
}
\arguments{
\item{mean}{The mean of the prior distribution.  An object of
class \code{\link[dembase:Values-class]{Values}}.}

\item{sd}{The standard deviation of the prior distribution.
If omitted, it is assumed to be 0.  \code{sd} can be an object
of class \code{\link[dembase:Values-class]{Values}}, or it can be
a single number, in which case it is applied to all
elements of \code{mean}.}
}
\value{
An object of class \code{\linkS4class{SpecKnown}}.
}
\description{
Specify a 'Known' prior.  The prior comes in two forms.  With the first
form, the corresponding main effect or interaction equals the mean
exactly,
\code{parameter[j] = mean[j]}.
With the second form,
\code{parameter[j] ~ N(mean[j], sd[j])}.
The second form is obtained by supplying a \code{sd} argument
to function \code{Known}.
}
\details{
Internally, the \code{mean} and \code{sd} arguments are permuted and
subsetted to make them compatible with the corresponding main effect or
interaction.  Thus, for example, if \code{mean} has values for regions
\code{A}, \code{B}, and \code{C}, but data \code{y} only includes
regions \code{A} and \code{B}, then \code{C} is silently ignored.

The original \code{mean} and \code{sd} arguments are, however,
stored, and can be used for prediction.  For example, we might
supply a \code{mean} argument with values for years 2005,
2010, 2015, 2020, and 2025, then call \code{\link{estimateModel}}
with data for 2005, 2010, and 2015, and finally call
\code{\link{predictModel}} to obtain forecasts for 2020 and 2025.
}
\examples{
mean <- ValuesOne(c(0.1, 0.2, 0.3),
                  labels = c("A", "B", "C"),
                  name = "region")
sd <- ValuesOne(c(0.02, 0.03, 0.02),
                  labels = c("A", "B", "C"),
                  name = "region")
## No uncertainty
Known(mean)

## Some uncertainty
Known(mean = mean, sd = sd)

## Single standard deviation
Known(mean = mean, sd = 0.05)
}
\seealso{
\code{ExchFixed} creates a normal prior centered at 0.
}

