% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Level}
\alias{Level}
\title{Specify the level term in a DLM prior.}
\usage{
Level(scale = HalfT())
}
\arguments{
\item{scale}{An object of class \code{\linkS4class{HalfT}}.}
}
\value{
An object of class \code{\linkS4class{Level}}
}
\description{
Specify the level term in a \code{\link{DLM}} prior.  Currently the only
part of the level term that can be specified is the prior for the
standard deviation of the innovations.
}
\examples{
Level()
Level(scale = HalfT(scale = 0.5))
}
\seealso{
\code{Level} is usually called as part of a call to
function \code{\link{DLM}}.
}
