% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{metropolis}
\alias{metropolis}
\title{Extract information on Metropolis-Hastings updates.}
\usage{
metropolis(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{SummaryResults}}.}
}
\value{
If Metropolis-Hastings updates where carried out, a matrix;
otherwise \code{NULL}.
}
\description{
Given an object of class \code{\linkS4class{SummaryResults}}, extract
the standard deviation of the proposal density, the proportion of
proposals accepted, and autocorrelation, for Metropolis-Hastings updates.
See the documentation for \code{\link{fetchSummary}} for details.
}
\examples{
library(demdata)
deaths <- Counts(round(VADeaths2))
popn <- Counts(VAPopn)
filename <- tempfile()
estimateModel(Model(y ~ Poisson(mean ~ age)),
              y = deaths,
              exposure = popn,
              filename = filename,
              nBurnin = 2,
              nSim = 5,
              nChain = 2)
summary.est <- fetchSummary(filename)
metropolis(summary.est)
}
\seealso{
\code{\link{fetchSummary}}
}

