% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Mix}
\alias{Mix}
\title{Specify a Mix prior.}
\usage{
Mix(along = NULL, components = Components(), weights = Weights(),
  covariates = NULL, error = Error(), maxComponents = 10)
}
\arguments{
\item{along}{Name of a dimension.}

\item{components}{An object of class \code{\linkS4class{Components}}.}

\item{weights}{An object of class \code{\linkS4class{Weights}}.}

\item{covariates}{An object of class
\code{\linkS4class{Covariates}} or \code{NULL}
(the default).  If \code{covariates} is
\code{NULL}, the prior does not include covariates.}

\item{error}{An object of class \code{\link{Error}}.}

\item{maxComponents}{The maximum number of components. Defaults to 10.}
}
\description{
A \code{Mix} prior is used to model an iteraction that includes
an "along" dimension (typically, but not necessarily, time),
and additional dimensions.  Values for the additional
dimensions evolve, generally smoothly, across successive values of the
"along" dimension.  A typical application is modelling age-structures
that change gradually over time.
}
\details{
The prior is modified from a model developed by Kunihama and Dunson
(2013).  The model is non-parametric, meaning that it extracts
structure from the data with minimal guidance from the user.  It is
designed for use with sparse data.
}
\examples{
Mix()
}
\references{
Kunihama, T and Dunson, DB. 2013. Bayesian modeling of temporal dependence in
large sparse contingency tables.  \emph{Journal of the American Statistical
Association}, 108(504): 1324-1338.
}
\seealso{
Priors for the components can be specified using
function \code{\link{Components}}, and weights (ie mixing parameters)
can be specified using function \code{\link{Weights}}.  However,
in normal usage, it is typically best to stick with the defaults.
}

