% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{Model}
\alias{Model}
\title{Specify a model for a single demographic series or
dataset.}
\usage{
Model(formula, ..., lower = NULL, upper = NULL, priorSD = NULL,
  jump = NULL, useHMC = NULL, sizeStep = NULL, nStep = NULL,
  series = NULL, aggregate = NULL)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} describing
the likelihood, and possibly parts of the  prior.
Constructed using functions such as \code{\link{Poisson}}.}

\item{\dots}{Non-default hyper-priors, in models that
include hyper-priors.  Constructed using functions
such as \code{\link{Exch}} and \code{\link{DLM}}.}

\item{lower}{A lower bound for estimates of data-level
means or probabilities (the \eqn{gamma_i}).}

\item{upper}{An upper bound for estimates of data-level
means or probabilities (the \eqn{gamma_i}).}

\item{priorSD}{An object of class \code{HalfT} specifying
the prior for the prior-level standard deviation.}

\item{jump}{The standard deviation of the proposal density
for Metropolis-Hasting updates.}

\item{useHMC}{Whether to update the main effect
and interaction terms in the prior model
using Hamiltonian Monte Carlo.
Defaults to \code{TRUE}. If \code{FALSE}, Gibbs
updates are used instead.}

\item{sizeStep}{The size of the steps used by the Hamiltonian
Monte Carlo updating of the prior model. Defaults to 0.1.}

\item{nStep}{The number of steps taken in each  Hamiltonian
Monte Carlo update of the prior model. Defaults to 10.}

\item{series}{The name of the demographic series (eg
\code{"population"} or \code{"births"}) that
is being modelled. Only needed when the model is to
be used in a call to \code{\link{estimateAccount}}.}

\item{aggregate}{An object of class \code{\linkS4class{SpecAggregate}}.}
}
\description{
The likelihood and, if the model has a second level,
main effects and interactions from that level, are
specified via functions such as \code{\link{Poisson}},
\code{\link{CMP}}, \code{\link{Binomial}},
\code{\link{Normal}}, or
\code{\link{PoissonBinomial}}.  \code{Model} is used
to specify the remaining parts of the model.
}
\details{
The \code{\dots} argument is used to specify non-default
hyper-priors.  Each main effect or interaction
specified via functions such as \code{\link{Poisson}}
has a default prior, which is determined from information such
as the {\link[dembase]{dimtype}} of the associated dimensions
of the series or dataset.  Since the characteristics of
the series or dataset are not known until functions
\code{\link{estimateModel}}, \code{\link{estimateCounts}},
or \code{\link{estimateAccount}} have been called, the
the process of choosing the default is not completed until
the \code{estimate} functions are called.

The \code{\dots} argument can be used to specify
non-default priors.  This is done by through expressions
such as

\code{region ~ Exch(error = Error(robust = TRUE))}

or

\code{region:time ~ DLM(trend = HalfT(scale = 0.2))}

where \code{region} and \code{time} are names of dimensions
in the dataset, and \code{\link{Exch}} and \code{\link{DLM}}
are functions for construction hyper-priors.

The \code{priorSD} argument refers to the standard
deviation at the second level of the model (if there
is a second level.)  It should not be confused with the
\code{priorsSD} argument to function \code{\link{Normal}},
which refers to the standard deviation in the likelihood.

The \code{lower} and \code{upper} arguments can be used
to constrain the range of the rate or count parameters
in the likelihood for Poisson models, the probability
parameters in the likelihood for binomial models,
or the mean parameters in the likelihood for normal models.
These constraints may reflect substantive features
of the application: for instance, in a normal model
it may make sense to constrain the means to be non-negative.
However, setting \code{lower} to a value slightly above 0
may also help resolve numerical problems in binomial
models were many estimated probabilities are close to 0,
and setting \code{upper} to a value slightly below 1 may
resolve problems when estimated probabilities are close to
1.

Printing the object created by \code{Model}, typically by typing
the name of the object at the console, shows the specification.
The \code{trunc-half-t(df, s^2, max)} in the printed results refers to a
truncated \code{\link[=halft-distn]{half-t}} distribution with \code{df}
degrees of freedom, scale \code{s^2}, and maximum value \code{max}.
}
\examples{
## model where all hyper-priors follow defaults
Model(y ~ Poisson(mean ~ age * sex + age * time))

## override default hyper-prior for age effect
Model(y ~ Poisson(mean ~ age * sex + age * time),
      age ~ DLM(trend = NULL))

## impose lower and upper bounds.
Model(y ~ Binomial(mean ~ age * region + time),
      lower = 0.001, upper = 0.999)

## increase size of Metropolis-Hastings steps
Model(y ~ Poisson(mean ~ age * sex + age * time),
      jump = 0.2)

## data model
Model(reg.births ~ PoissonBinomial(prob = 0.98),
      series = "births")

overall.av <- AgNormal(value = 0.3, sd = 0.01)
Model(y ~ Binomial(mean ~ region + sex),
      aggregate = overall.av)
}
