% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Norm-class}
\alias{Norm-class}
\title{An S4 class to specify a normal distribution.}
\description{
An object of class \code{Norm} specifies a normal
distribution.  It is used as a prior distribution.
}
\section{Slots}{

\describe{
\item{\code{mean}}{Mean.}

\item{\code{A}}{Standard deviation.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Norm} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{Norm} are created
    by calls to function \code{\link{Norm}}.
}
