% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{NormalFixed}
\alias{NormalFixed}
\title{Specify a model based on a normal distribution with known
means and standard deviations.}
\usage{
NormalFixed(mean, sd, useExpose = TRUE)
}
\arguments{
\item{mean}{An object of class \code{\link[dembase:Values-class]{Values}}
holding means.}

\item{sd}{The standard deviation of the means. \code{sd} can be an object
of class \code{\link[dembase:Values-class]{Values}}, or it can be
a single number, in which case it is applied to all
elements of \code{mean}.}

\item{useExpose}{Whether the model includes an
exposure term. Defaults to \code{TRUE}.}
}
\value{
An object of class \code{\linkS4class{SpecLikelihoodNormalFixed}}.
}
\description{
Specify a model of the form
  \deqn{y_i = N(mean[i], sd[i])}
or
  \deqn{y_i = N(exposure[i] * mean[i], sqrt(exposure[i]) * sd[i])}.
}
\details{
Among other things, the model is useful as a data model for
surveys.  In such cases, \code{exposure} represents the
true underlying counts, \code{mean} represents coverage rates,
\code{sd} represents standard errors for the coverage rates,
and \code{y} represents the observered counts.  If the model
is being used to represent a census post-enumeration survey,
for instance, then \code{exposure} would be the true population,
\code{mean} the coverage rates as estimated from the survey,
\code{se} the survey-based standard errors, and \code{y} the
census counts.

Subscript \eqn{i} denotes a cell within a classification
defined by variables such as age, sex, and time.
For instance cell \eqn{i} might be 30-34 year old females
in 2020.
}
\examples{
mean <- Values(array(c(0.95, 0.96, 0.95, 0.94),
                     dim = c(2, 2),
                     dimnames = list(age = c("0-39", "40+"),
                                     sex = c("Female", "Male"))))
sd <- Values(array(c(0.05, 0.08, 0.05, 0.04),
                   dim = c(2, 2),
                   dimnames = list(age = c("0-39", "40+"),
                                   sex = c("Female", "Male"))))
NormalFixed(mean = mean, sd = sd)
NormalFixed(mean = mean, sd = 0.1)
NormalFixed(mean = mean, sd = 0.1, useExpose = FALSE)
}
\seealso{
\code{NormalFixed} is typically used as
part of a call to function \code{\link{Model}}.
More flexible normal models can be specified using
\code{\link{Normal}}.
}
