% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{parameters}
\alias{parameters}
\title{Extract summaries of parameter estimates from a SummaryResults object.}
\usage{
parameters(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{SummaryResults}}.}
}
\value{
A data.frame.
}
\description{
Given an object of class \code{\linkS4class{SummaryResults}}, extract
a data.frame containing summaries of parameter estimates.
See the documentation for \code{\link{fetchSummary}} for details
of the summaries.
}
\examples{
deaths <- demdata::VADeaths2
popn <- demdata::VAPopn
deaths <- round(deaths)
deaths <- Counts(deaths)
popn <- Counts(popn)
filename <- tempfile()
model <- Model(y ~ Poisson(mean ~ age + sex),
               jump = 0.5)
estimateModel(model = model,
              y = deaths,
              exposure = popn,
              filename = filename,
              nBurnin = 50,
              nSim = 50,
              nChain = 2,
              parallel = FALSE)
summary.est <- fetchSummary(filename)
parameters(summary.est)
}
\seealso{
\code{\link{fetchSummary}}
}
