% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plotHalfT}
\alias{plotHalfT}
\title{Plot the half-t distribution.}
\usage{
plotHalfT(df = 7, scale = 1, max = 0.999, density = TRUE,
  add = FALSE, ...)
}
\arguments{
\item{df}{Degrees of freedom. Positive. Can be non-integer,
and can be \code{Inf}.}

\item{scale}{Dispersion parameter.}

\item{max}{A quantile, defaulting to 0.999.  The x-axis for the plot
extends from 0 to this quantile.}

\item{density}{Whether to plot the density function (the default) or the
distribution function.}

\item{add}{Whether to add to the current plot.}

\item{\dots}{Other arguments, passed to functions \code{\link[graphics]{plot}}
and \code{\link[graphics]{lines}}.}
}
\description{
Plot the density or distribution function of a
\code{\link[=halft-distn]{half-t}} distribution.
}
\examples{
plotHalfT()
plotHalfT(df = 4, add = TRUE, col = "red")
plotHalfT(df = 4, scale = 1.1, add = TRUE, col = "blue")
}
