% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{PoissonBinomial}
\alias{PoissonBinomial}
\title{Specify a model based on a Poisson-binomial mixture.}
\usage{
PoissonBinomial(prob)
}
\arguments{
\item{prob}{The probability that a person or event
is correctly enumerated. A number between 0 and 1,
and typically close to 1.}
}
\value{
An object of class \code{\linkS4class{SpecLikelihood}}.
}
\description{
Specify a model of the form
  \deqn{y_i = U_i + V_i}
  \deqn{U_i \sim Poisson((1-p) n_i)}
  \deqn{V_i \sim binomial(n_i, p)}.
}
\details{
The model is useful mainly as a way of representing
the relationship between a true set of counts (the
\eqn{n_i}) and measurements of those counts (the
\eqn{y_i}.)  For instance, \eqn{n_i} could be true
counts of births, and \eqn{y_i} could be births
recorded in an accurate births registration system.

Subscript \eqn{i} denotes a cell within a classification
defined by variables such as age, sex, and time.
For instance cell \eqn{i} might be 30-34 year old females
in 2020.

Higher values of \eqn{p} imply greater accuracy;
\eqn{p} can be interpreted as the probability that
a person or event is enumerated in the correct
cell \eqn{i}.

One limitation of the model is that it does not allow
for the possibility that \eqn{y_i > 0} when
\eqn{n_i = 0}. In other words, it does not allow for the
possibility that a person or event is erroneously
enumerated in a cell that has no people or events.
}
\examples{
PoissonBinomial(prob = 0.98)
}
\seealso{
\code{PoissonBinomial} is typically used as
part of a call to function \code{\link{Model}}.
}

