% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-functions.R
\name{predictAccount}
\alias{predictAccount}
\title{Use results from function estimateAccount to make predictions.}
\usage{
predictAccount(filenameEst, filenamePred, along = NULL, labels = NULL,
  n = NULL, data = NULL, aggregate = NULL, lower = NULL, upper = NULL,
  nBurnin = 0L, parallel = TRUE, verbose = FALSE, useC = TRUE)
}
\arguments{
\item{filenameEst}{Filename used to \code{\link{estimateModel}}.}

\item{filenamePred}{Filename to be used by \code{\link{predictModel}}.}

\item{along}{Name or index along which prediction should occur.}

\item{labels}{Labels of new categories.}

\item{n}{Number of new categories.  Can only be used when predicting
along \code{"time"}, \code{"age"}, or \code{"cohort"} dimensions,
and when the units have equal length.}

\item{data}{A named list containing future values of covariates.}

\item{aggregate}{An object of class \code{\linkS4class{SpecAggregate}}.}

\item{lower}{A lower bound for estimates of data-level rates,
probabilities, or means.}

\item{upper}{An upper bound for estimates of data-level
rate, probabilities, or means.}

\item{nBurnin}{Number of iteration discarded before recording begins.}

\item{parallel}{Logical.  If \code{TRUE} (the default), parallel processing
is used.}

\item{verbose}{Logical.  If \code{TRUE} (the default) a message is
printed at the end of the calculations.}

\item{useC}{Logical.  If \code{TRUE} (the default), the calculations
are done in C.  Setting \code{useC} to \code{FALSE} may be useful
for debugging.}
}
\description{
Not written yet.
}
