% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{Round3}
\alias{Round3}
\title{Specify a data model for random rounding to base 3.}
\usage{
Round3()
}
\value{
An object of class \code{\linkS4class{SpecLikelihood}}.
}
\description{
Specify a model in which the rounded value \code{x[i]} is obtained from the
original value \code{q[i]} using the rule described in
\code{\link[dembase:Values-class]{Values}}.
}
\details{
The model is useful for analysing data that have been
confidentialised to base 3. It is used as a data model
in calls to \code{\link{estimateCount}}, and
\code{\link{estimateAccount}}.
}
\examples{
Round3()
}
\seealso{
\code{Round3} is typically used as
part of a call to function \code{\link{Model}}.
}
