% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Season-class}
\alias{Season-class}
\title{An S4 class to specify a seasonal effect in a DLM prior.}
\description{
A prior specified using \code{\link{DLM}}
can include a seasonal effect.  The innovation
term for the seasonal effect has a normal distribution,
the standard deviation for which has a truncated
half-\emph{t} distribution.
}
\section{Slots}{

\describe{
\item{\code{nSeason}}{Number of seasons.}

\item{\code{nuSeason}}{Degrees of freedom of truncated half-t
prior for standard deviation.}

\item{\code{ASeason}}{Scale of truncated half-t
prior for standard deviation.}

\item{\code{omegaSeasonMax}}{Maximum value of truncated half-t
prior for standard deviation.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Season} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Objects of class \code{Season} are created
    using function \code{\link{Season}}.
}

