% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Season}
\alias{Season}
\title{Specify a seasonal effect in a DLM prior.}
\usage{
Season(n, scale = HalfT())
}
\arguments{
\item{n}{Number of seasons.  An integer greater than 1.}

\item{scale}{Prior for the \code{scaleSeason} parameter. An object of class
\code{\linkS4class{HalfT}}.}
}
\value{
An object of class \code{\linkS4class{Season}}
}
\description{
A \code{\link{DLM}} prior can contain a seasonal effect, which
is specified using function \code{Season}. Season effects can
can change over time.
}
\details{
In a prior for a main effect, the seasonal effect has the form

\code{s[j] = s[j-n] + errorSeason[j],}

where \code{n} is the number of seasons.

In a prior for an interaction, the seasonal effect has the form

\code{s[k,l] = s[k-n,l] + errorSeason[k,l].}

(See the documentation for function \code{\link{DLM}} for
an explanation of the \code{k,l} subscripts.)

The error term has the form

\code{errorSeason[j] ~ N(0, scaleSeason^2)}

or

\code{errorSeason[k,l] ~ N(0, scaleSeason^2)}.

Standard deviation \code{scaleSeason} has a half-\emph{t} prior,
which can be specified using function \code{\link{HalfT}}.
Season effects are not fixed, but instead evolve over time.  The speed
with which season effects change is governed by \code{scaleSeason}.
}
\examples{
Season(n = 4)
Season(n = 2, scale = HalfT(scale = 3))
}
\seealso{
\code{Season} supplies an option argument
to function \code{\link{DLM}}.
}

