% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecDLM-class}
\alias{SpecDLM-class}
\alias{SpecDLMNoTrendNormCovNoSeason-class}
\alias{SpecDLMNoTrendNormCovWithSeason-class}
\alias{SpecDLMNoTrendNormZeroNoSeason-class}
\alias{SpecDLMNoTrendNormZeroWithSeason-class}
\alias{SpecDLMNoTrendRobustCovNoSeason-class}
\alias{SpecDLMNoTrendRobustCovWithSeason-class}
\alias{SpecDLMNoTrendRobustZeroNoSeason-class}
\alias{SpecDLMNoTrendRobustZeroWithSeason-class}
\alias{SpecDLMWithTrendNormCovNoSeason-class}
\alias{SpecDLMWithTrendNormCovWithSeason-class}
\alias{SpecDLMWithTrendNormZeroNoSeason-class}
\alias{SpecDLMWithTrendNormZeroWithSeason-class}
\alias{SpecDLMWithTrendRobustCovNoSeason-class}
\alias{SpecDLMWithTrendRobustCovWithSeason-class}
\alias{SpecDLMWithTrendRobustZeroNoSeason-class}
\alias{SpecDLMWithTrendRobustZeroWithSeason-class}
\title{An S4 class to specify a dynamic linear model (DLM) prior.}
\description{
An object of class \code{SpecDLM} specifies a prior
in which elements that are next to each other are
are expected to be more similar than elements that
are distant for each other.  DLMs are typically used to
model variation over time, but are often also appropriate
for variation over age.
}
\details{
The prior consists of a level term, optional trend,
seasonal, and covariates terms, and an error term.
The error term can follow a normal or t distribution.
For details, see the documentation for function
\code{\link{DLM}}.
}
\section{Slots}{

\describe{
\item{\code{along}}{The name of the dimension that the
time series (or age equivalents) extend along.}

\item{\code{phi}}{Damping parameter.}

\item{\code{phiKnown}}{Whether the damping parameter is known
or is estimated from the data.}

\item{\code{minPhi}}{Minimum value for the damping parameter.}

\item{\code{maxPhi}}{Maximum value for the damping parameter.}

\item{\code{nuAlpha}}{Degrees of freedom for the prior for
the standard deviation for innovations
in the level term.}

\item{\code{AAlpha}}{Scale for the prior for the standard
deviation for innovations in the level term.}

\item{\code{omegaAlphaAMax}}{Maximum for the standard deviation
for innovations in the level term.}

\item{\code{nuDelta}}{Degrees of freedom for the prior for
the standard deviation for innovations
in the trend term.}

\item{\code{ADelta}}{Scale for the prior for the standard
deviation for innovations in the trend term.}

\item{\code{omegaDeltaAMax}}{Maximum for the standard deviation
for innovations in the trend term.}

\item{\code{nSeason}}{Number of seasons.}

\item{\code{nuSeason}}{Degrees of freedom for the prior for
the standard deviation for innovations
in the season term.}

\item{\code{ASeason}}{Scale for the prior for the standard
deviation for innovations in the season term.}

\item{\code{omegaSeasonAMax}}{Maximum for the standard deviation
for innovations in the season term.}

\item{\code{formula}}{A \code{\link[stats]{formula}} with response
\code{mean}.}

\item{\code{data}}{A data.frame containing covariate data.}

\item{\code{contrastsArg}}{A named list, the elements which are matrices
or names of contrasts functions.}

\item{\code{AIntercept}}{The standard deviation for the prior
for the intercept in the covariates term.}

\item{\code{nuEtaCoef}}{The degrees of freedom for the prior
for the coefficients in the covariates term.}

\item{\code{AEtaCoef}}{The scale for the prior
for the coefficients in the covariates term.}

\item{\code{nuBeta}}{The degrees of freedom for the error
term, if the error term has a t distribution.}

\item{\code{nuTau}}{The degrees of freedom for the truncated
half-t prior for the standard deviation or scale
parameter for the error term.}

\item{\code{ATau}}{The scale for the truncated
half-t prior for the standard deviation or scale
parameter for the error term.}

\item{\code{maxTau}}{The maximum value for the standard
deviation or scale parameter for the error term.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecDLM} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Object of class \code{SpecDLM} are generated
using function \code{\link{DLM}}.
}

