% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecExch-class}
\alias{SpecExch-class}
\alias{SpecExchNormCov-class}
\alias{SpecExchNormZero-class}
\alias{SpecExchRobustCov-class}
\alias{SpecExchRobustZero-class}
\title{A S4 class to specify an exchangeable prior.}
\description{
An object of class \code{SpecExch} specifies
a prior in which the order of the elements is
irrelevant.  The prior consists of an optional
covariates term, and an error term. The error term
can have a normal or t distribution. For details,
see the documentation for function \code{Exch}.
}
\section{Slots}{

\describe{
\item{\code{formula}}{A \code{\link[stats]{formula}} with response
\code{mean}.}

\item{\code{data}}{A data.frame containing covariate data.}

\item{\code{contrastsArg}}{A named list, the elements which are matrices
or names of contrasts functions.}

\item{\code{AIntercept}}{The standard deviation for the prior
for the intercept in the covariates term.}

\item{\code{nuEtaCoef}}{The degrees of freedom for the prior
for the coefficients in the covariates term.}

\item{\code{AEtaCoef}}{The scale for the prior
for the coefficients in the covariates term.}

\item{\code{nuBeta}}{The degrees of freedom for the error
term, if the error term has a t distribution.}

\item{\code{nuTau}}{The degrees of freedom for the truncated
half-t prior for the standard deviation or scale
parameter for the error term.}

\item{\code{ATau}}{The scale for the truncated
half-t prior for the standard deviation or scale
parameter for the error term.}

\item{\code{maxTau}}{The maximum value for the standard
deviation or scale parameter for the error term.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecExch} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Object of class \code{SpecExch} are generated
using function \code{\link{Exch}}.
}

