% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecKnown-class}
\alias{SpecKnown-class}
\title{A S4 class to specify a Known prior.}
\description{
An object of class \code{SpecExchFixed} specifies
a prior in which
  parameter[j] ~ N(mean[j], sd[j]^2),
where sd[j] can be 0. For details, see the documentation
for function \code{\link{Known}}.
}
\section{Slots}{

\describe{
\item{\code{alphaKnown}}{Mean values.}

\item{\code{AKnownVec}}{Standard deviations.}

\item{\code{metadata}}{Metadata for mean values and standard deviations.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecKnown} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Object of class \code{SpecExchFixed} should be generated
using function \code{\link{ExchFixed}}.  More complicated
exchangeable priors are specified by object of class
\code{\linkS4class{SpecExch}}.
}
