% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-14-Model-hierarchy-spec.R
\docType{class}
\name{SpecLikelihood-class}
\alias{SpecLikelihood-class}
\alias{SpecLikelihoodBinomial-class}
\alias{SpecLikelihoodCMP-class}
\alias{SpecLikelihoodNormalVarsigmaKnown-class}
\alias{SpecLikelihoodNormalVarsigmaUnknown-class}
\alias{SpecLikelihoodPoisson-class}
\alias{SpecLikelihoodPoissonBinomialMixture-class}
\alias{SpecLikelihoodNormalFixed-class}
\alias{SpecLikelihoodRound3-class}
\title{S4 classes to specify one or two levels of a model.}
\description{
Classes to specify the likelihood and, if the model
has higher levels, part or all of the second level.
}
\section{Slots}{

\describe{
\item{\code{formulaMu}}{Object of class \code{\link[stats]{formula}}
specifying main effects and interactions in second level
of model.}

\item{\code{varsigma}}{Data-level standard deviation,
when this is supplied by the user (and treated
as known.)}

\item{\code{nuVarsigma}}{Degrees of freedom for truncated
half-t prior for data-level standard deviation.}

\item{\code{AVarsigma}}{Scale for truncated
half-t prior for data-level standard deviation.}

\item{\code{varsigmaMax}}{Maximum value for data-level
standard deviation.}

\item{\code{mean}}{Vector of means.}

\item{\code{sd}}{Vector of standard deviations.}

\item{\code{metadata}}{Metadata for \code{mean} and \code{sd}.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecLikelihood} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{SpecLikelihood}
are created by calls to function such as
\code{\link{Poisson}}, \code{\link{Binomial}},
\code{\link{Normal}}, and \code{\link{PoissonBinomial}}.
}
