% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecMix-class}
\alias{SpecMix-class}
\title{An S4 class to specify a Mix prior.}
\description{
An object of class \code{SpecMix} specifies a non-parametric
prior, based on a mix of normal distributions, that can
be used for modelling interactions, particularly interactions
involving time.
}
\section{Slots}{

\describe{
\item{\code{along}}{The name of the "along" dimension.}

\item{\code{indexClassMaxMix}}{The maximum number of components that
can be accommodated.}

\item{\code{minLevelComponentWeight}}{The lower bound for values of
\code{levelComponentWeightMix}.}

\item{\code{maxLevelComponentWeight}}{The upper bound for values of
\code{levelComponentWeightMix}.}

\item{\code{multComponentWeightMix}}{Multiplier applied to the default
value for \code{omegaComponentWeightMix}.}

\item{\code{multLevelComponentWeightMix}}{Multiplier applied to the default
value for \code{omegaLevelComponentWeightMix}.}

\item{\code{multVectorsMix}}{Multiplier applied to the default
value for \code{omegaVectorsMix}.}

\item{\code{nuComponentWeightMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaComponentWeightMix}.}

\item{\code{nuLevelComponentWeightMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaLevelComponentWeightMix}.}

\item{\code{nuVectorsMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaVectorsMix}.}

\item{\code{priorMeanLevelComponentWeightMix}}{Mean for normal prior
for \code{meanLevelComponentWeightMix}.}

\item{\code{priorSDLevelComponentWeightMix}}{Standard deviation for normal prior
for \code{meanLevelComponentWeightMix}}

\item{\code{AComponentWeightMix}}{Scale for truncated half-t prior for
\code{omegaComponentWeightMix}.}

\item{\code{ALevelComponentWeightMix}}{Scale for truncated half-t prior for
\code{omegaLevelComponentWeightMix}.}

\item{\code{AVectorsMix}}{Scale for truncated half-t prior for
\code{omegaVectorsMix}.}

\item{\code{omegaComponentWeightMaxMix}}{Upper limit for truncated half-t prior
for \code{omegaComponentWeightMaxMix}.}

\item{\code{omegaLevelComponentWeightMaxMix}}{Upper limit for truncated half-t prior
for \code{omegalevelComponentWeightMaxMix}.}

\item{\code{omegaVectorsMix}}{Upper limit for truncated half-t prior
for \code{omegaVectorsMix}.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecMix} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Objects of class \code{SpecMix} are generated using function
\code{\link{Mix}}.
}

