% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-14-Model-hierarchy-spec.R
\docType{class}
\name{SpecModel-class}
\alias{SpecModel-class}
\alias{SpecBinomialVarying-class}
\alias{SpecNormalVaryingVarsigmaUnknown-class}
\alias{SpecNormalVaryingVarsigmaKnown-class}
\alias{SpecPoissonVarying-class}
\alias{SpecPoissonBinomialMixture-class}
\alias{SpecNormalFixed-class}
\title{S4 classes to specify a model.}
\description{
Classes describing all the parts of a model that can
be specified wihout knowing the exact structure of,
and, possibly variation in, the data.
}
\details{
The specification might, for instance, contain a list
of main effects and interactions to be included
in a hierarchical model, but not the priors for
themain effects and interactions,  which depend
on the length and \code{\link[dembase]{dimtype}}
of the dimensions.  These are not known until the
relevant data, such as outcome variable \code{y},
are supplied, which occurs in the call to function
\code{\link{estimateModel}}, \code{\link{estimateCounts}},
or \code{\link{estimateAccount}}.
}
\section{Slots}{

\describe{
\item{\code{call}}{The original \code{\link[base]{call}}
to function \code{\link{Model}}.}

\item{\code{nameY}}{The name of the outcome variable,
which, in models for the data, may be the name
of a dataset.}

\item{\code{series}}{The name of the demographic series
being modelled.  Used only when dealing with demographic
accounts.}

\item{\code{varsigma}}{Data-level standard deviation,
when this is supplied by the user (and treated
as known.)}

\item{\code{nuVarsigma}}{Degrees of freedom for truncated
half-t prior for data-level standard deviation.}

\item{\code{AVarsigma}}{Scale for truncated
half-t prior for data-level standard deviation.}

\item{\code{varsigmaMax}}{Maximum value for data-level
standard deviation.}

\item{\code{prob}}{In a Poisson-binomial model, the probability
that a person or event is enumerated and is placed in
the correct cell.}

\item{\code{lower}}{Lower limit for the data-level rate,
probability, or mean parameter.}

\item{\code{upper}}{Upper limit for the data-level rate,
probability, or mean parameter.}

\item{\code{tolerance}}{Small quantity added to \code{lower}
or subtracted from \code{upper} when testing
whether a proposed value for a data-level rate,
probability, or mean is within the required bounds.}

\item{\code{maxAttempt}}{Maximum number of attempts at
generating a proposal for a data-level rate,
probability or mean before giving up and retaining
the current value, within one iteration of the
Gibbs sampler.}

\item{\code{scaleTheta}}{The standard deviation of the
proposal density for Metropolis-Hastings updates
of the data-level rate, probability, or mean
parameter.}

\item{\code{formulaMu}}{A \code{\link[stats]{formula}}
describing the main effects and interactions in
a hierarchical model.}

\item{\code{specsPriors}}{A list of object of class
\code{\linkS4class{SpecPrior}}, describing
any non-default priors for main effects and
interactions.}

\item{\code{nameSpecPriors}}{The names of the main effects
or interactions that have non-default priors.}

\item{\code{nuSigma}}{Degrees of freedom for truncated
half-t prior for standard deviation in prior
(level 2) model.}

\item{\code{ASigma}}{Scale for truncated
half-t prior for standard deviation in prior
(level 2) model.}

\item{\code{sigmaMax}}{Maximum value for standard deviation
in prior (level 2) model.}

\item{\code{aggregate}}{An object of class
\code{\linkS4class{SpecAggregate}}.}

\item{\code{mean}}{Vector of means in NormalFixed model - subsetted to
include only cells that are observed in \code{y}.}

\item{\code{meanAll}}{Vector of means in NormalFixed model, before
subsetting.}

\item{\code{sd}}{Vector of standard deviations in NormalFixed model - subsetted to
include only cells that are observed in \code{y}}

\item{\code{sdAll}}{Vector of standard deviations in NormalFixed model, before
subsetting.}

\item{\code{metadata}}{Metadata for \code{mean} and \code{sd}.}

\item{\code{metadataAll}}{Metadata for \code{meanAll} and \code{sdAll}.}

\item{\code{useExpose}}{Whether the model includes and exposure term.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{SpecModel} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Object of class \code{SpecModel} are created
by a call to function \code{\link{Model}}.
}
