% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecPrior-class}
\alias{SpecPrior-class}
\title{A S4 superclass for prior specifications.}
\description{
All classes that describe priors inherit from
class \code{SpecPrior}.  Examples include
\code{\linkS4class{SpecExchFixed}}, \code{\linkS4class{SpecExch}}
\code{\linkS4class{SpecDLM}}, and \code{\linkS4class{SpecMix}}.
}
