% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-08-Prior-hierarchy-spec.R
\docType{class}
\name{SpecZero-class}
\alias{SpecZero-class}
\title{A S4 class to specify a Zero prior.}
\description{
Specify a prior that fixes all values of the associated
main effect or interaction to zero.
}
\details{
The class has no slots.
}
\seealso{
An object of class \code{SpecZero} should be generated
using function \code{\link{Zero}}.
}

