% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-21-Summary.R, R/Summary-methods.R
\docType{class}
\name{SummaryResults-class}
\alias{SummaryResults-class}
\alias{show,SummaryDataset-method}
\alias{show,SummaryModel-method}
\alias{show,SummaryResultsCounts-method}
\alias{show,SummaryResultsModelEst-method}
\alias{show,SummaryResultsModelPred-method}
\alias{show,SummarySeries-method}
\title{S4 class summarizing results from estimation or prediction.}
\usage{
\S4method{show}{SummaryModel}(object)

\S4method{show}{SummaryDataset}(object)

\S4method{show}{SummarySeries}(object)

\S4method{show}{SummaryResultsModelEst}(object)

\S4method{show}{SummaryResultsModelPred}(object)

\S4method{show}{SummaryResultsCounts}(object)
}
\arguments{
\item{object}{Object of class \code{SummaryResults}.}
}
\description{
\code{SummaryResults} is the superclass for classes holding
results from functions \code{\link{estimateModel}},
\code{\link{estimateCounts}}, \code{\link{estimateAccount}},
\code{\link{predictModel}}, \code{\link{predictCounts}}, and
\code{\link{predictAccount}} (though the summaries from the
predict functions are generally less informative.)
}
\details{
Objects of class \code{SummaryResults} are created using
function \code{\link{fetchSummary}}.  Individual components from
summary objects can be extracted using functions \code{\link{gelmanDiag}},
\code{\link{metropolis}}, and \code{\link{parameters}}.
}
\section{Slots}{

\describe{
\item{\code{mcmc}}{named integer vector containing nBurnin, nSim, nChain,
nThin, nIteration}

\item{\code{parameters}}{data.frame summarizing parameter estimates.}
}}

