% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{TDist-class}
\alias{TDist-class}
\title{An S4 class to a vector of independent t-distributed variables}
\description{
An object of class \code{TDist} is used to specify the
prior distribution of coefficents in the \code{\link{Covariates}}
part of a prior prior.
}
\section{Slots}{

\describe{
\item{\code{nu}}{Degrees of freedom}

\item{\code{mean}}{Means}

\item{\code{A}}{Scales}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{TDist} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{TDist} are created
by calls to function \code{\link{TDist}}.
}
