% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{TFixed}
\alias{TFixed}
\title{Specify a model based on a Student's t distribution with known
location, scale, and degrees of freedom.}
\usage{
TFixed(location, scale, df = 7, useExpose = TRUE)
}
\arguments{
\item{location}{An object of class \code{\link[dembase:Values-class]{Values}}
specifying the center of the distribution.}

\item{scale}{Parameter governing dispersion. Can be an object
of class \code{\link[dembase:Values-class]{Values}}, or a single number,
a single number, in which case it is applied to all
elements of \code{location}.}

\item{df}{A positive number. Defaults to 7.}

\item{useExpose}{Whether the model includes an
exposure term. Defaults to \code{TRUE}.}
}
\value{
An object of class \code{\linkS4class{SpecLikelihoodTFixed}}.
}
\description{
Specify a model of the form
  \deqn{y_i = t(location[i], scale[i], df)}
or
  \deqn{y_i = t(exposure[i] * location[i], scale[i], df)}.
}
\details{
The model is useful as a data model where a few cells
have much larger errors than the rest. Smaller values for \code{df}
lead to thicker tails (ie larger probabilities of extreme values.)
Subscript \eqn{i} denotes a cell within a classification
defined by variables such as age, sex, and time.
For instance cell \eqn{i} might be 30-34 year old females
in 2020.
}
\examples{
location <- Values(array(c(0.95, 0.96, 0.95, 0.94),
                     dim = c(2, 2),
                     dimnames = list(age = c("0-39", "40+"),
                                     sex = c("Female", "Male"))))
scale <- Values(array(c(0.05, 0.08, 0.05, 0.04),
                   dim = c(2, 2),
                   dimnames = list(age = c("0-39", "40+"),
                                   sex = c("Female", "Male"))))
TFixed(location = location, scale = scale)
TFixed(location = location, scale = 0.1, df = 1)
TFixed(location = location, scale = 0.1, useExpose = FALSE)

## indistinguishable from normal distribution:
TFixed(location = location, scale = scale, df = 30)

## Cauchy distribution (very heavy tails):
TFixed(location = location, scale = scale, df = 1)
}
\seealso{
\code{TFixed} is typically used as
part of a call to function \code{\link{Model}}.
Similar data models that do not allow for the possibility of occasional
large errors can be specified using \code{\link{Normal}} (though
setting \code{df} to a value greater than 30 in \code{TFixed} will
have much the same effect.)
Flexible normal models can be specified using
\code{\link{Normal}}.
}
