% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Trend}
\alias{Trend}
\title{Specify the trend term in a DLM prior.}
\usage{
Trend(initial = Initial(), scale = HalfT())
}
\arguments{
\item{initial}{An object of class \code{\linkS4class{Initial}}.}

\item{scale}{An object of class \code{\linkS4class{HalfT}}.}
}
\value{
An object of class \code{\linkS4class{Trend}}
}
\description{
Specify the prior for the initial values, and the prior for the standard
deviation of the innovations in a \code{\link{DLM}} prior.
}
\examples{
Trend()
Trend(initial = Initial(mean = 0.05, sd = 0.1),
      scale = HalfT(scale = 0.5))
}
\seealso{
\code{Trend} is usually called as part of a call to
function \code{\link{DLM}}.
}

