% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Weights-class}
\alias{Weights-class}
\title{An object of class \code{Weights} is used to specify the
trend term in a \code{\link{Mix}} prior.}
\description{
An object of class \code{Weights} is used to specify the
trend term in a \code{\link{Mix}} prior.
}
\section{Slots}{

\describe{
\item{\code{minLevelComponentWeight}}{The lower bound for values of
\code{levelComponentWeightMix}.}

\item{\code{maxLevelComponentWeight}}{The upper bound for values of
\code{levelComponentWeightMix}.}

\item{\code{multComponentWeightMix}}{Multiplier applied to the default
value for \code{omegaComponentWeightMix}.}

\item{\code{multLevelComponentWeightMix}}{Multiplier applied to the default
value for \code{omegaLevelComponentWeightMix}.}

\item{\code{nuComponentWeightMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaComponentWeightMix}.}

\item{\code{nuLevelComponentWeightMix}}{The degrees of freedom for the truncated
half-t prior for \code{omegaLevelComponentWeightMix}.}

\item{\code{priorMeanLevelComponentWeightMix}}{Mean for normal prior
for \code{meanLevelComponentWeightMix}.}

\item{\code{priorSDLevelComponentWeightMix}}{Standard deviation for normal prior
for \code{eanLevelComponentWeightMix}}

\item{\code{AComponentWeightMix}}{Scale for truncated half-t prior for
\code{omegaComponentWeightMix}.}

\item{\code{ALevelComponentWeightMix}}{Scale for truncated half-t prior for
\code{omegaLevelComponentWeightMix}.}

\item{\code{omegaComponentWeightMaxMix}}{Upper limit for truncated half-t prior
for \code{omegaComponentWeightMaxMix}.}

\item{\code{omegaLevelComponentWeightMaxMix}}{Upper limit for truncated half-t prior
for \code{omegalevelComponentWeightMaxMix}.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Weights} object.  The slots are not part of
the API of the package, and may change in future.
}

