% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Weights}
\alias{Weights}
\title{Specify priors for the weights in a Mix prior.}
\usage{
Weights(mean = 0, sd = 1, damp = Damp(), scale1 = HalfT(),
  scale2 = HalfT())
}
\arguments{
\item{mean}{The mean of the prior for \code{mean}. Defaults to 0.}

\item{sd}{The standard deviation of the prior for \code{mean}.
Defaults to 1.}

\item{damp}{An object of class \code{\linkS4class{Damp}}.}

\item{scale1}{An object of class \code{\linkS4class{HalfT}} defining
the prior for \code{error1}.}

\item{scale2}{An object of class \code{\linkS4class{HalfT}} defining
the prior for \code{error2}.}
}
\value{
An object of class \code{\linkS4class{Weights}}.
}
\description{
In normal usage, it is better to accept the default priors
for the weights than to specify them explicitly.
End users are therefore unlikely to need this function.
}
\details{
A \code{\link{Mix}} prior treats an interaction as a mixture of
normal distributions.  The weights, also referred as mixture
parameters, evolve over time.  The evolution of the weight for
each component is governed by an auto-regressive time series model.
Specifically, transformed values of the weights are modelled
using

\code{level1[k,h] = level2[k,h] + error1[k,h]}

\code{level2[k,h] = mean + damp * level2[k-1,h] + error2[k,h]}

\code{mean} has a normal prior.

\code{damp} has the boundary-avoiding prior described in
Gelman et al (2004) p316-317.  This prior is equivalent to assuming
a Beta(2, 2) prior on the transformed parameter (\code{damp}+0.5)/2.

\code{error1} and \code{error2} both have half-t priors. These
priors have the defaults described in \code{\link{HalfT}}.
}
\seealso{
Mixture priors are specified using \code{\link{Mix}}.
The weights in a mixture prior are specified using
\code{\link{Weights}}.
}
