% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Zero}
\alias{Zero}
\title{Specify a prior that sets all terms to zero.}
\usage{
Zero()
}
\arguments{
\item{mean}{The mean of the prior distribution.  An object of}
}
\value{
An object of class \code{\linkS4class{Zero}}.
}
\description{
Specify a prior for a main effect or interaction in which 
\code{parameter[j] = 0}.
}
\details{
In some cases, as with a dimension representing upper and lower
Lexis triangles, setting all elements of the main effect or interaction
to zero may make substantive sense.  In other cases, a \code{Zero} prior
to reduce the number of terms being estimated, while conforming
to that requirement from \code{\link{Model}} that all terms marginal
to an interaction be included in a model.
}
\examples{
Zero()
}
\seealso{
\code{Known} can be used to create a prior that
fixes the main effect or interaction at values other than 0.
}
