% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-14-Model-hierarchy-spec.R
\docType{class}
\name{Dispersion-class}
\alias{Dispersion-class}
\title{S4 class to specify prior for dispersion parameter in CMP model.}
\description{
Specify the mean and standard deviation of (the log of) the
dispersion parameter ('nu') in a CMP model.  Also specify
the prior mean and standard deviation of the mean.
}
\section{Slots}{

\describe{
\item{\code{meanLogNuCMP}}{The prior mean of the dispersion parameter.}

\item{\code{sdLogNuCMP}}{The prior standard deviation of the
dispersion parameter.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Dispersion} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{Dispersion} are created
by calls to function \code{\link{Dispersion}}.
}
