% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Level-class}
\alias{Level-class}
\title{An S4 class to specify the level term in a DLM prior.}
\description{
An object of class \code{Level} is used to specify the
level term in a \code{\link{DLM}} prior.
}
\section{Slots}{

\describe{
\item{\code{nuAlpha}}{Degrees of freedom.}

\item{\code{AAlpha}}{Scale.}

\item{\code{omegaAlphaMax}}{Maximum value for standard deviation of
innovations.}

\item{\code{multAlpha}}{Multiplier applied to default value for \code{AAlpha}.}
}}
\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Level} object.  The slots are not part of
the API of the package, and may change in future.
}
\seealso{
Objects of class \code{Level} are created
by calls to function \code{\link{Level}}.
}

