% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecModel-generators.R
\name{likelihood}
\alias{likelihood}
\alias{Binomial}
\alias{Normal}
\alias{Poisson}
\alias{CMP}
\title{Specify first two levels of hierarchical model.}
\usage{
Poisson(formula, useExpose = TRUE, structuralZeros = NULL, boxcox = 0)

CMP(
  formula,
  dispersion = Dispersion(),
  useExpose = TRUE,
  structuralZeros = NULL,
  boxcox = 0
)

Binomial(formula)

Normal(formula, sd = NULL, priorSD = HalfT())
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} with response
\code{mean} and names of dimensions of demographic
series or dataset being modelled.}

\item{useExpose}{Whether the model includes an
exposure term. Defaults to \code{TRUE}.}

\item{structuralZeros}{Location of any structural zeros
in the data. An object of class \code{\link[dembase:Values-class]{Values}},
or, for the typical case, the word \code{"diag"}.}

\item{boxcox}{Parameter determining transformation of rates
or counts used in Poisson or CMP models. Defaults to 0,
implying that a log transform is used.}

\item{dispersion}{The dispersion parameter for a CMP model.
An object of class \code{\linkS4class{Dispersion}},
typically created by a call to function \code{\link{Dispersion}}.}

\item{sd}{Standard deviation in the likelihood for the
normal model.  If a value is not supplied,
it is estimated from the data.}

\item{priorSD}{An object of class \code{HalfT} specifying
a non-default prior for the standard deviation in the
likelihood for the normal model.}
}
\value{
An object of class \code{\linkS4class{SpecLikelihood}}.
}
\description{
Specify the likelihood and part of the prior for a
Poisson, binomial, or normal hierarchical model.
}
\details{
Specify a likelihood and prior of the form
\deqn{y_i \sim Poisson(\gamma_i n_i)}
\deqn{log(\gamma_i) \sim N(x_i \beta, \sigma_i),}

\deqn{y_i \sim Poisson(\gamma_i)}
\deqn{log(\gamma_i) \sim N(x_i \beta, \sigma_i),}

\deqn{y_i \sim CMP(\gamma_i n_i, \nu_i)}
\deqn{log(\gamma_i) \sim N(x_i \beta, \sigma_i),}
\deqn{log(\nu_i) \sim N(m, s^2)}

\deqn{y_i \sim CMP(\gamma_i, \nu_i)}
\deqn{log(\gamma_i) \sim N(x_i \beta, \sigma_i),}
\deqn{log(\nu_i) \sim N(m, s^2)}

\deqn{y_i \sim binomial(n_i, \gamma_i)}
\deqn{logit(\gamma_i) \sim N(x_i \beta, \sigma_i),}
or
\deqn{y_i \sim normal(\gamma_i, \phi^2 / w_i)}
\deqn{\gamma_i \sim N(x_i \beta, \sigma_i)}.

Subscript \eqn{i} denotes a cell within a multiway array,
such as an array with dimensions age, sex, and time.  In
Poisson and binomial models, \eqn{y_i} is a count.
The \eqn{n_i} term is an exposure in the case of Poisson
and CMP (COMPoisson) models and a
sample size in the case of binomial models.
It is not supplied in calls to function \code{Poisson}
or \code{Binomial}.  In normal models \eqn{y_i}
is a cell-specific value such as a mean,
and \eqn{w_i} is a cell-specific weight.  Weights
are not supplied in calls to function \code{Normal}.
Vector \eqn{\beta} contains main effects
and interactions, such as age effects, time effects,
and age-region interactions. Vector \eqn{x_i} is the
\eqn{i}th row from the design matrix \eqn{X}.

The main effects and interactions are specified
via the \code{formula} argument. For instance,

  \code{mean ~ age * sex + time}

specifies a model with age, sex, and time main effects,
and an age-sex interaction.

The main effects and interactions in a hierarchical model are only weakly
identified: see the documentation for function \code{\link{fetch}} for
details.

If a model has two or more levels, the second level
typically contains more than just main effects
and interactions.  For instance, the second level of
Poisson, binomial, and normal hierarchical models
contains a variance term.  The remaining parts of
the second level, such as the variances, as well
as any higher levels, are specified
in calls to function \code{\link{Model}}, or to
functions \code{\link{estimateModel}},
\code{\link{estimateCounts}}, or \code{\link{estimateAccount}}.

Poisson and CMP models allow structural zeros in the data,
that is, cells whose value must be zero by definition.
Examples include the number of pregnant males or the
number of people transitioning straight from "single"
to "divorced". The most general way to specify structural zeros
is to supply an object of class \code{\link[dembase:Values-class]{Values}},
with zeros in the places where structural zeros are expected,
and non-zero values elsewhere. The Values object only has to
contain enough dimensions to specify the positions where the structural
zeros appear: other dimensions will be added as needed.  See below
for an example.

The most common situation where structural zeros
are required is when the data has origin and destination dimensions,
and values on the diagonal are always zero.  Setting
\code{structuralZeros} to \code{"diag"} is equivalent to
supplying a Values object with zeros on the diagonal.
}
\examples{
## age effect, sex effect, age-sex interaction,
## and time effect
Poisson(mean ~ age * sex + time)

## same model, but without exposure term
Poisson(mean ~ age * sex + time, useExpose = FALSE)

## use formula notation to specify second-order interactions
Binomial(mean ~ (age + sex + region)^2)

Normal(mean ~ age + education + income)
## specify the exact value of the standard deviation
Normal(mean ~ age + education + income,
       sd = 0.3)
## specify a non-default prior for the standard deviation
Normal(mean ~ age + education + income,
       priorSD = HalfT(scale = 100))

## Specify structural zero on the diagonal
struc.zeros <- Values(array(c(0, 1, 1,
                              1, 0, 1,
                              1, 1, 0),
                            dim = c(3, 3),
                            dimnames = list(region_orig = c("A", "B", "C"),
                                          region_dest = c("A", "B", "C"))))
## Note that the model contains age and sex dimensions. These
## The pattern of zeros specified by 'struc.zeros' will be
## replicated for each combination of these dimensions.
Poisson(mean ~ region_orig * region_dest + age * sex,
        structuralZeros = struc.zeros)
## The same pattern of structural zeros, with zeros on the diagonal
## formed by the origina and destination dimensions, can be specified
## by using the word "diag"
Poisson(mean ~ region_orig * region_dest + age * sex,
        structuralZeros = "diag")
}
\seealso{
Functions \code{Poisson}, \code{Binomial}, and \code{Normal} are
used as part of a call to function \code{\link{Model}}.
}
