% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Norm}
\alias{Norm}
\title{Specify a normal distribution.}
\usage{
Norm(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean. Defaults to 0.}

\item{sd}{Standard deviation. A positive number. Defaults to 1.}
}
\value{
An object of class \code{\linkS4class{Norm}}.
}
\description{
Specify a normal distribution. \code{Norm} is used to specify
a prior (on the log scale) for
the dispersion parameter in a \code{\link{CMP}} model.
}
\examples{
Norm()
Norm(sd = 0.5)
Norm(mean = 0.5, sd = 1.5)
}
\seealso{
\code{Norm} is used in calls to function \code{\link{Covariates}}.
}
