% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Norm}
\alias{Norm}
\title{Specify a normal distribution centered at 0.}
\usage{
Norm(scale = NULL)
}
\arguments{
\item{scale}{Standard deviation for the normal distribution.
A positive number.}
}
\value{
An object of class \code{\linkS4class{Norm}}.
}
\description{
Specify a normal distribution with mean equal to 0 and standard
deviation equal to \code{scale}.
}
\details{
\code{Norm} is used to specify the prior for the intercept in
function \code{\link{Covariates}}.  If a value for \code{scale}
is not specified, a default value is determined when function
\code{\link{estimateModel}}, \code{\link{estimateCounts}}, or
\code{\link{estimateAccount}} is called.  Let \eqn{s} be the standard
deviation of data \eqn{y}, or of \code{log(y)} in the case of a Poisson
model without exposure.  Then the default for \code{scale} is
determined as follows.

\tabular{ll}{
  \strong{Model} \tab \strong{Default} \cr
  Poisson with exposure \tab \eqn{10} \cr
  Poisson without exposure \tab  \eqn{10s} \cr
  binomial \tab \eqn{10} \cr
  normal \tab  \eqn{10s}
}
}
\examples{
Norm()
Norm(scale = 0.5)
}
\seealso{
\code{Norm} is used in calls to function \code{\link{Covariates}}.
}
