% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-functions.R
\name{simulateModel}
\alias{simulateModel}
\title{Draw parameters and data from a statistical model}
\usage{
simulateModel(model, y = NULL, exposure = NULL, weights = NULL,
  filename = NULL, nDraw = 10, verbose = TRUE, useC = TRUE)
}
\arguments{
\item{model}{An object of class \code{\linkS4class{SpecModel}},
specifying the model to be fit.}

\item{y}{A \code{\link[dembase:DemographicArray-class]{demographic array}}
holding the outcome data.}

\item{exposure}{A \code{\link[dembase:DemographicArray-class]{Counts}}
object specifying exposure or sample size.}

\item{weights}{A \code{\link[dembase:DemographicArray-class]{Counts}}
object containing weights.}

\item{filename}{The name of a file where output is collected.}

\item{nDraw}{The number of random draws to make from the model.}

\item{verbose}{Logical.  If \code{TRUE} (the default) a message is
printed at the end of the calculations.}

\item{useC}{Logical.  If \code{TRUE} (the default), the calculations
are done in C.  Setting \code{useC} to \code{FALSE} may be useful
for debugging.}

\item{nCore}{The number of cores to use if \code{parallel} is
\code{TRUE}. Defaults to 4.}
}
\description{
WARNING - THIS FUNCTION IS STILL UNDER DEVELOPMENT
}
