% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-11-Aggregate.R
\docType{class}
\name{SpecAggregate-class}
\alias{SpecAggregate-class}
\alias{SpecAgCertain-class}
\alias{SpecAgUncertain-class}
\alias{SpecAgNormal-class}
\alias{SpecAgFun-class}
\alias{SpecAgPoisson-class}
\alias{SpecAgLife-class}
\title{S4 classes to represent aggregate values.}
\description{
Aggregate values can be used to represent additional
information beyond what is contained in 'y',
such as expert judgements.
}
\section{Slots}{

\describe{
\item{\code{metadataAg}}{Metadata (such as \code{\link[dembase]{dimtypes}}
for aggregate values.}

\item{\code{scaleAg}}{Standard deviation of proposal density
for Metropolis-Hastings updates.}

\item{\code{valueAg}}{Parameter(s) estimated by the aggregate values.}

\item{\code{funAg}}{Function to calculate elements of \code{valueAg} from
disaggregated rates, counts, probabilities, or means.}

\item{\code{weightAg}}{Weights used to calculate \code{valueAg} from
disaggregated rates, counts, probabilities, or means.}

\item{\code{concordancesAg}}{List of \code{\link[classconc]{ManyToOne}}
concordances.}
}}

