% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-10-Prior-generators.R
\docType{class}
\name{Trend-class}
\alias{Trend-class}
\title{An S4 class to specify the trend term in a DLM prior.}
\description{
An object of class \code{Trend} is used to specify the
trend term in a \code{\link{DLM}} prior.
}
\section{Slots}{

\describe{
\item{\code{ADelta}}{Scale for half-t prior for standard deviation
of innovations.}

\item{\code{ADelta0}}{Standard deviation of the normal prior
for the first trend term.}

\item{\code{meanDelta0}}{Mean of the normal prior for the first
trend term. Defaults to 0.}

\item{\code{multDelta}}{Multiplier applied to code{ADelta}.}

\item{\code{multDelta0}}{Multiplier applied to code{ADelta0}.}

\item{\code{nuDelta}}{Degrees of freedom for half-t prior for
standard deviation of innovations.}

\item{\code{omegaDeltaMax}}{Maximum value for standard deviation
of innovations.}
}}

\section{Warning}{

In  normal usage, it should not be necessary to
access, or even know about, the slots of a
\code{Trend} object.  The slots are not part of
the API of the package, and may change in future.
}

\seealso{
Objects of class \code{Trend} are created
by calls to function \code{\link{Trend}}.
}
