% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPrior-generators.R
\name{Weights}
\alias{Weights}
\title{Specify priors for the weights in a Mix prior.}
\usage{
Weights(mean = 0, sd = 1, scale1AR = HalfT(), scale2AR = HalfT(),
  minAR2 = -4, maxAR2 = 4)
}
\arguments{
\item{mean}{The mean of the prior for \code{meanAR}. Defaults to 0.}

\item{sd}{The standard deviation of the prior for \code{meanAR}.
Defaults to 1.}

\item{minAR2}{Minimum value that \code{levelAR2} can take. Defaults to -4.}

\item{maxAR2}{Maximum value that \code{levelAR2} can take. Defaults to 4.}

\item{scaleAR1}{An object of class \code{\linkS4class{HalfT}} defining
the prior for \code{errorAR1}.}

\item{scaleAR2}{An object of class \code{\linkS4class{HalfT}} defining
the prior for \code{errorAR2}.}
}
\value{
An object of class \code{\linkS4class{Weights}}.
}
\description{
In normal usage, it is better to accept the default priors
for the weights than to specify them explicitly.
End users are therefore unlikely to need this function.
}
\details{
A \code{\link{Mix}} prior treats an interaction as a mixture of
normal distributions.  The weights, also referred as mixture
parameters, evolve over time.  The evolution of the weight for
each component is governed by an auto-regressive time series model.
Specifically, transformed values of the weights are modelled
using

\code{levelAR1[k,h] = levelAR2[k,h] + errorAR1[k,h]}

\code{levelAR2[k,h] = meanAR + dampAR * level2AR[k-1,h] + error2AR[k,h]}

\code{meanAR} has a normal prior.

\code{dampAR} has the boundary-avoiding prior described in
Gelman et al (2004) p316-317.  This prior is equivalent to assuming
a Beta(2, 2) prior on the transformed parameter (\code{dampAR}+0.5)/2.

\code{errorAR1} and \code{errorAR2} both have half-t priors. These
priors have the defaults described in \code{\link{HalfT}}.

To avoid numerical problems, \code{levelAR2} is confined to the
interval (\code{minAR2}, \code{maxAR2}). Users can override the
default values, including setting them to \code{-Inf} and \code{Inf}.
USER-SUPPLIED VALUES CURRENTLY IGNORED.
}
\seealso{
Mixture priors are specified using \code{\link{Mix}}.
The weights in a mixture prior are specified using
\code{\link{Weights}}.
}
